% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation_linter.R
\name{add_indentation_data}
\alias{add_indentation_data}
\title{Add indentation data to the parsed content}
\usage{
add_indentation_data(pc)
}
\arguments{
\item{pc}{parsed_content as produced by \code{get_source_file}}
}
\value{
a modified parsed_content data.frame with additional fields
  relating to line indentation, function header indentation and indentation
  relative to a parent expression's line.
}
\description{
Add indentation data to the parsed content
}
