% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{add_linter}
\alias{add_linter}
\title{Add a Linter}
\usage{
add_linter(name, linter)
}
\arguments{
\item{name}{The name of the linter, as a string.}

\item{linter}{A \code{\link{linter}}.}
}
\description{
Add a linter, to be used in subsequent calls to \code{\link{lint}}.
}

