% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{all_linters}
\alias{all_linters}
\title{Create a linter configuration based on all available linters}
\usage{
all_linters(packages = "lintr")
}
\arguments{
\item{packages}{A character vector of packages to search for linters.}
}
\description{
Create a linter configuration based on all available linters
}
\examples{
names(all_linters())

}
\seealso{
\itemize{
\item \link{linters_with_defaults} for basing off lintr's set of default linters.
\item \link{linters_with_tags} for basing off tags attached to linters, possibly across multiple packages.
\item \link{available_linters} to get a data frame of available linters.
\item \link{linters} for a complete list of linters available in lintr.
}
}
