% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_wrapping_linter.R
\name{argument_wrapping_linter}
\alias{argument_wrapping_linter}
\title{Checks function header parameters:}
\usage{
argument_wrapping_linter(
  source_file,
  expr_xml_node,
  length = 80L,
  indent = {     2L }
)
}
\description{
# Lint (function header overflows an 80 character line limit):
  my_func <- function(args, causing, the, function, header, to, overflow, the, linewidth) {
  }
}
\details{
# Lint-free:
  my_func <- function(args, causing, the, function, header, to, overflow, the,
    linewidth) {
  }


  # Lint (a multi-line parameter default is mixed with spanning arguments):
  my_func <- function(one, multiline = {
      1L,
    }, arg) {
  }

  # Lint-free:
  my_func <- function(
    one,
    multiline = {
      1L
    },
    args) {
  }


Checks function call arguments

  # Lint (overflows 80 character line limit):
  vapply(mtcars, function(i) { i["mpg"] / i["wt"] }, numeric(1L), USE.NAMES = FALSE)

  # Lint-free:
  vapply(mtcars, function(i) { i["mpg"] / i["wt"] }, numeric(1L),
    USE.NAMES = FALSE)


  # Lint (a multi-line argument is mixed with spanning arguments):
  lapply(data, function(i) {
      i
    })

  # Lint-free:
  lapply(
    data,
    function(i) {
      i
    })
}
