% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_eol_linter.R
\name{assignment_eol_linter}
\alias{assignment_eol_linter}
\title{Assignment End of Line linter}
\usage{
assignment_eol_linter(allow_comments = FALSE)
}
\arguments{
\item{allow_comments}{Suppress lints for commented out lines}
}
\description{
Check that there are no assignment operators at the end of source lines.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=style_linters]{style}
}
