% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_linter.R
\name{assignment_linter}
\alias{assignment_linter}
\title{Assignment linter}
\usage{
assignment_linter(
  allow_cascading_assign = TRUE,
  allow_right_assign = FALSE,
  allow_trailing = TRUE
)
}
\arguments{
\item{allow_cascading_assign}{Logical, default \code{TRUE}.
If \code{FALSE}, \code{\link[base:assignOps]{<<-}} and \verb{->>} are not allowed.}

\item{allow_right_assign}{Logical, default \code{FALSE}. If \code{TRUE}, \verb{->} and \verb{->>} are allowed.}

\item{allow_trailing}{Logical, default \code{TRUE}. If \code{FALSE} then assignments aren't allowed at end of lines.}
}
\description{
Check that \verb{<-} is always used for assignment.
}
\examples{
# will produce lints
lint(
  text = "x = mean(x)",
  linters = assignment_linter()
)

lint(
  text = "1 -> x; 2 ->> y",
  linters = assignment_linter()
)

# okay
lint(
  text = "x <- mean(x)",
  linters = assignment_linter()
)

lint(
  text = "x <- 1; y <<- 2",
  linters = assignment_linter()
)

# customizing using arguments
lint(
  text = "1 -> x; 2 ->> y",
  linters = assignment_linter(allow_right_assign = TRUE)
)

lint(
  text = "x <<- 1",
  linters = assignment_linter(allow_cascading_assign = FALSE)
)

writeLines("foo(bar = \n 1)")
lint(
  text = "foo(bar = \n 1)",
  linters = assignment_linter(allow_trailing = FALSE)
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr. \cr
\url{https://style.tidyverse.org/syntax.html#assignment-1}
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
