% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_spaces_linter.R
\name{assignment_spaces_linter}
\alias{assignment_spaces_linter}
\title{Assignment spaces linter}
\usage{
assignment_spaces_linter()
}
\description{
Check that assignments only have one space before and after.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=readability_linters]{readability}, \link[=style_linters]{style}
}
