% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean_arithmetic_linter.R
\name{boolean_arithmetic_linter}
\alias{boolean_arithmetic_linter}
\title{Require usage of boolean operators over equivalent arithmetic}
\usage{
boolean_arithmetic_linter()
}
\description{
\code{length(which(x == y)) == 0} is the same as \code{!any(x == y)}, but the latter
is more readable and more efficient.
}
\details{
#' @evalRd rd_tags("boolean_arithmetic_linter")
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
