% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_dataframe_linter.R
\name{cbind_dataframe_linter}
\alias{cbind_dataframe_linter}
\title{Recommend direct usage of \code{data.frame()} to create a data.frame from a list}
\usage{
cbind_dataframe_linter()
}
\description{
It is possible to create a data.frame from a list of columns with \code{data.frame()}
or \code{list2DF()} (since \R 4.0.0 and if recycling is not required), rather than
iteratively adding columns with \code{cbind()}.
}
\examples{
# will produce lints
lint(
  text = "do.call(cbind.data.frame, x)",
  linters = cbind_dataframe_linter()
)

# okay
lint(
  text = "list2DF(x)",
  linters = cbind_dataframe_linter()
)

lint(
  text = "data.frame(list(letters, LETTERS))",
  linters = cbind_dataframe_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}
}
