% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_only_linter.R
\name{character_only_linter}
\alias{character_only_linter}
\title{Enforce library calls using symbols}
\usage{
character_only_linter()
}
\description{
R packages should be imported as symbols, like \code{library(lintr)},
and \emph{not} via strings like \code{library("lintr", character.only = TRUE)},
with few exceptions.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
