% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{checkstyle_output}
\alias{checkstyle_output}
\title{Checkstyle Report for lint results}
\usage{
checkstyle_output(lints, filename = "lintr_results.xml")
}
\arguments{
\item{lints}{the linting results.}

\item{filename}{the name of the output report}
}
\description{
Generate a report of the linting results using the
\href{https://checkstyle.sourceforge.io/}{Checkstyle} XML format.
}
