% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the lintr cache}
\usage{
clear_cache(file = NULL, path = NULL)
}
\arguments{
\item{file}{filename whose cache to clear. If you pass \code{NULL}, it will delete all of the caches.}

\item{path}{directory to store caches. Reads option 'lintr.cache_directory' as the default.}
}
\value{
0 for success, 1 for failure, invisibly.
}
\description{
Clear the lintr cache
}
