% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_curly_linter.R
\name{closed_curly_linter}
\alias{closed_curly_linter}
\title{Closed curly linter}
\usage{
closed_curly_linter(allow_single_line = FALSE)
}
\arguments{
\item{allow_single_line}{if \code{TRUE}, allow an open and closed curly pair on the same line.}
}
\description{
Check that closed curly braces are on their own line unless they follow an else.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
