% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_linters.R
\name{comment_linter}
\alias{comment_linter}
\title{Comment linter}
\usage{
comment_linter()
}
\description{
Check that comments follow proper style.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
