% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{common_mistakes_linters}
\alias{common_mistakes_linters}
\title{Common mistake linters}
\description{
Linters highlighting common mistakes, such as duplicate arguments.
}
\section{Linters}{
The following linters are tagged with 'common_mistakes':
\itemize{
\item{\code{\link{duplicate_argument_linter}}}
\item{\code{\link{equals_na_linter}}}
\item{\code{\link{missing_argument_linter}}}
\item{\code{\link{missing_package_linter}}}
\item{\code{\link{sprintf_linter}}}
}
}
