% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_message_linter.R
\name{condition_message_linter}
\alias{condition_message_linter}
\title{Block usage of \code{paste()} and \code{paste0()} with messaging functions using \code{...}}
\usage{
condition_message_linter()
}
\description{
This is because
\itemize{
\item \code{stop(paste0(...))} is redundant because it is exactly equivalent to \code{stop(...)}
\item \code{stop(paste(...))} is also equivalent to \code{stop(...)} with separators (see examples)
}
}
\details{
The same applies to all default condition functions, i.e., \code{\link[=stop]{stop()}}, \code{\link[=warning]{warning()}},
\code{\link[=message]{message()}}, and \code{\link[=packageStartupMessage]{packageStartupMessage()}}.
}
\examples{
# will produce lints
lint(
  text = "stop(paste('a string', 'another'))",
  linters = condition_message_linter()
)

lint(
  text = "warning(paste0('a string', ' another'))",
  linters = condition_message_linter()
)

# okay
lint(
  text = "stop('a string', ' another')",
  linters = condition_message_linter()
)

lint(
  text = "warning('a string', ' another')",
  linters = condition_message_linter()
)

lint(
  text = "warning(paste('a string', 'another', sep = '-'))",
  linters = condition_message_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}
}
