% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunct_expectation_linter.R
\name{conjunct_expectation_linter}
\alias{conjunct_expectation_linter}
\title{Force && conditions in expect_true(), expect_false() to be written separately}
\usage{
conjunct_expectation_linter()
}
\description{
For readability of test outputs, testing only one thing per call to
\code{\link[testthat:logical-expectations]{testthat::expect_true()}} is preferable, i.e.,
\verb{expect_true(A); expect_true(B)} is better than \code{expect_true(A && B)}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
No tags are given.
}
