% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_stopifnot_linter.R
\name{consecutive_stopifnot_linter}
\alias{consecutive_stopifnot_linter}
\title{Force consecutive calls to stopifnot into just one when possible}
\usage{
consecutive_stopifnot_linter()
}
\description{
\code{stopifnot()} accepts any number of tests, so sequences like
\verb{stopifnot(x); stopifnot(y)} are redundant.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
