% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_suppression_linter.R
\name{consecutive_suppression_linter}
\alias{consecutive_suppression_linter}
\title{Force consecutive calls suppress* calls into just one when possible}
\usage{
consecutive_suppression_linter()
}
\description{
Suppressing the output of consecutive calls to \code{\link[=library]{library()}} with
\code{\link[=suppressMessages]{suppressMessages()}} or \code{\link[=suppressPackageStartupMessages]{suppressPackageStartupMessages()}} need only be
done once, by wrapping the sequence of library calls into
one \code{{}} expression, as opposed to repeatedly
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
