% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{consistency_linters}
\alias{consistency_linters}
\title{Consistency linters}
\description{
Linters checking enforcing a consistent alternative if there are multiple syntactically valid ways to write
something.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'consistency':
\itemize{
\item{\code{\link{assignment_linter}}}
\item{\code{\link{class_equals_linter}}}
\item{\code{\link{comparison_negation_linter}}}
\item{\code{\link{condition_message_linter}}}
\item{\code{\link{consecutive_assertion_linter}}}
\item{\code{\link{consecutive_mutate_linter}}}
\item{\code{\link{function_argument_linter}}}
\item{\code{\link{if_not_else_linter}}}
\item{\code{\link{if_switch_linter}}}
\item{\code{\link{implicit_integer_linter}}}
\item{\code{\link{inner_combine_linter}}}
\item{\code{\link{is_numeric_linter}}}
\item{\code{\link{keyword_quote_linter}}}
\item{\code{\link{length_levels_linter}}}
\item{\code{\link{literal_coercion_linter}}}
\item{\code{\link{nested_pipe_linter}}}
\item{\code{\link{nrow_subset_linter}}}
\item{\code{\link{null_coalescing_linter}}}
\item{\code{\link{numeric_leading_zero_linter}}}
\item{\code{\link{nzchar_linter}}}
\item{\code{\link{object_name_linter}}}
\item{\code{\link{paste_linter}}}
\item{\code{\link{print_linter}}}
\item{\code{\link{quotes_linter}}}
\item{\code{\link{redundant_ifelse_linter}}}
\item{\code{\link{rep_len_linter}}}
\item{\code{\link{scalar_in_linter}}}
\item{\code{\link{seq_linter}}}
\item{\code{\link{system_file_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
\item{\code{\link{unnecessary_nesting_linter}}}
\item{\code{\link{which_grepl_linter}}}
\item{\code{\link{whitespace_linter}}}
}
}
