% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse1_linter.R
\name{deparse1_linter}
\alias{deparse1_linter}
\title{Require usage of deparse1() over paste(deparse(x), collapse = ...)}
\usage{
deparse1_linter()
}
\description{
\code{paste(deparse(x), collapse = ...)} is the same as \code{deparse1(x)}, but
harder to read.
}
\examples{
# will produce lints
lint(
  text = 'paste(deparse(x), collapse = " ")',
  linters = deparse1_linter()
)

lint(
  text = 'paste0(deparse(x), collapse = "")',
  linters = deparse1_linter()
)

# okay
lint(
  text = "deparse1(x)",
  linters = deparse1_linter()
)

lint(
  text = "deparse(x)",
  linters = deparse1_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
