% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{efficiency_linters}
\alias{efficiency_linters}
\title{Efficiency linters}
\description{
Linters highlighting code efficiency problems, such as unnecessary function calls.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'efficiency':
\itemize{
\item{\code{\link{any_duplicated_linter}}}
\item{\code{\link{any_is_na_linter}}}
\item{\code{\link{boolean_arithmetic_linter}}}
\item{\code{\link{consecutive_mutate_linter}}}
\item{\code{\link{fixed_regex_linter}}}
\item{\code{\link{if_switch_linter}}}
\item{\code{\link{ifelse_censor_linter}}}
\item{\code{\link{inner_combine_linter}}}
\item{\code{\link{leap_year_linter}}}
\item{\code{\link{length_test_linter}}}
\item{\code{\link{lengths_linter}}}
\item{\code{\link{list2df_linter}}}
\item{\code{\link{literal_coercion_linter}}}
\item{\code{\link{matrix_apply_linter}}}
\item{\code{\link{nested_ifelse_linter}}}
\item{\code{\link{nrow_subset_linter}}}
\item{\code{\link{nzchar_linter}}}
\item{\code{\link{outer_negation_linter}}}
\item{\code{\link{redundant_equals_linter}}}
\item{\code{\link{redundant_ifelse_linter}}}
\item{\code{\link{regex_subset_linter}}}
\item{\code{\link{routine_registration_linter}}}
\item{\code{\link{sample_int_linter}}}
\item{\code{\link{scalar_in_linter}}}
\item{\code{\link{seq_linter}}}
\item{\code{\link{sort_linter}}}
\item{\code{\link{string_boundary_linter}}}
\item{\code{\link{unnecessary_concatenation_linter}}}
\item{\code{\link{unnecessary_lambda_linter}}}
\item{\code{\link{vector_logic_linter}}}
\item{\code{\link{which_grepl_linter}}}
}
}
