% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/else_same_line_linter.R
\name{else_same_line_linter}
\alias{else_same_line_linter}
\title{Require else to come on the same line as \}, if present}
\usage{
else_same_line_linter()
}
\description{
This linter catches \code{if}/\verb{else} clauses where \code{if} uses \verb{\\\{} and its terminal
\verb{\\\}} is on a different line than the matched \verb{else}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=readability_linters]{readability}, \link[=style_linters]{style}
}
