% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Exclude lines or files from linting}
\usage{
exclude(lints, exclusions = settings$exclusions, linter_names = NULL, ...)
}
\arguments{
\item{lints}{that need to be filtered.}

\item{exclusions}{manually specified exclusions}

\item{linter_names}{character vector of names of the active linters, used for parsing inline exclusions.}

\item{...}{additional arguments passed to \code{\link[=parse_exclusions]{parse_exclusions()}}}
}
\description{
Exclude lines or files from linting
}
\details{
Exclusions can be specified in three different ways.
\enumerate{
\item Single line in the source file. default: \verb{# nolint}, possibly followed by a listing of linters to exclude.
If the listing is missing, all linters are excluded on that line. The default listing format is
\verb{# nolint: linter_name, linter2_name.}. There may not be anything between the colon and the line exclusion tag
and the listing must be terminated with a full stop (\code{.}) for the linter list to be respected.
\item Line range in the source file. default: \verb{# nolint start}, \verb{# nolint end}. \verb{# nolint start} accepts linter
lists in the same form as \verb{# nolint}.
\item Exclusions parameter, a list named and/or unnamed entries. If present, the name is a path relative to
the config. Moreover, elements have the following characteristics:
\enumerate{
\item Unnamed elements specify filenames or directories.
\item Named elements are numeric or a list. If present, the name is a linter.
\enumerate{
\item Numeric or unnamed list elements are line numbers to be excluded.
\item Named list elements are line numbers to be excluded for specific linters.
}
}
}
}
