% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_wrapper_linter.R
\name{exclude_wrapper_linter}
\alias{exclude_wrapper_linter}
\title{Exclude linters from execution on files}
\usage{
exclude_wrapper_linter(
  ...,
  exclusions = NULL,
  pattern = rex::rex(".", one_of("Rr"), or("", "html", "md", "nw", "rst", "tex", "txt"),
    end),
  root = getwd()
)
}
\arguments{
\item{...}{all linters that need to be exclude with extra rules}

\item{exclusions}{exclusions path, relative to the package path.}

\item{pattern}{pattern for files, by default it will take files with any of the extensions
.R, .Rmd, .qmd, .Rnw, .Rhtml, .Rrst, .Rtex, .Rtxt allowing for lowercase r (.r, ...).}

\item{root}{Base directory for relative filename resolution.}
}
\description{
This linters wrapper is very useful when creating linters
for lint_dir or lint_package
}
\examples{

# will produce lints
lint(
  text = strrep("x", 23L),
  linters = exclude_wrapper_linter(line_length_linter(20), exclusions = c("inst/", "vignettes/"))
)

lint(
  text = paste0(strrep("x", 23L), "\nany(!x))"),
  linters = exclude_wrapper_linter(
    line_length_linter(20), outer_negation_linter(),
    exclusions = c("inst/", "vignettes/")
  )
)

# okay
lint(
  text = strrep("x", 23L),
  linters = exclude_wrapper_linter(
    line_length_linter(20),
    exclusions = c(normalizePath(tempdir(), mustWork = TRUE)),
    pattern = "^file"
   )
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=package_development_linters]{package_development}
}
