% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_lint.R
\name{expect_app_lint_free}
\alias{expect_app_lint_free}
\title{Test that the shiny app is lint free}
\usage{
expect_app_lint_free(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{lint_shinyapp}}}
}
\description{
This function is a thin wrapper around lint_shinyapp that simply tests there are no
lints in the app.
}
