% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_error_linter.R
\name{expect_error_linter}
\alias{expect_error_linter}
\title{Require usage of expect_error(x) where appropriate}
\usage{
expect_error_linter()
}
\description{
If code is expected to produce an error, using \code{expect_that(x, throws_error())} works but is unadvised. It is better to use the function
dedicated specifically for checking errors: \code{expect_error(x)}. The additional
benefit is that you can also check for the expected error message using
regular expressions.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=package_development_linters]{package_development}
}
