% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_length_linter.R
\name{expect_length_linter}
\alias{expect_length_linter}
\title{Require usage of \code{expect_length(x, n)} over \code{expect_equal(length(x), n)}}
\usage{
expect_length_linter()
}
\description{
\code{\link[testthat:expect_length]{testthat::expect_length()}} exists specifically for testing the \code{\link[=length]{length()}} of
an object. \code{\link[testthat:equality-expectations]{testthat::expect_equal()}} can also be used for such tests,
but it is better to use the tailored function instead.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=package_development_linters]{package_development}, \link[=readability_linters]{readability}
}
