% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expect_lint.R
\name{expect_lint_free}
\alias{expect_lint_free}
\title{Test that the given content is lint free, this is
equivalent to \code{expect_lint(content, NULL, ..., file)}.}
\usage{
expect_lint_free(content, ..., file = NULL)
}
\arguments{
\item{content}{the file content to be linted}

\item{...}{one or more linters to use for the check}

\item{file}{if not \code{NULL} read content from a file rather than from \code{content}}
}
\description{
Test that the given content is lint free, this is
equivalent to \code{expect_lint(content, NULL, ..., file)}.
}

