% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_setequal_linter.R
\name{expect_setequal_linter}
\alias{expect_setequal_linter}
\title{Require usage of expect_setequal() over expect_identical(sort(...), ...)}
\usage{
expect_setequal_linter()
}
\description{
\code{expect_setequal()} is designed for testing an output, regardless of order;
for example, this is particularly useful for SQL, which is typically
row-order-agnostic.
}
\details{
Note that, in the presence of possible duplicates,
\code{expect_identical(sort(x), y)} won't be the same as
\code{expect_setequal(x, y)}. This linter encourages a separate test for
duplicates rather than integrating two tests into one.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=package_development_linters]{package_development}, \link[=readability_linters]{readability}
}
