% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_shape_linter.R
\name{expect_shape_linter}
\alias{expect_shape_linter}
\title{Require usage of \code{expect_shape(x, ...)} over \code{expect_equal(nrow(x), n)} and similar}
\usage{
expect_shape_linter()
}
\description{
\code{\link[testthat:expect_shape]{testthat::expect_shape()}} exists specifically for testing the \code{\link[=nrow]{nrow()}}, \code{\link[=ncol]{ncol()}},
or \code{\link[=dim]{dim()}} of an object. \code{\link[testthat:expect_equal]{testthat::expect_equal()}} and
\code{\link[testthat:expect_identical]{testthat::expect_identical()}} can also be used for such tests,
but it is better to use the tailored function instead.
}
\examples{
# will produce lints
lint(
  text = "expect_equal(nrow(x), 4L)",
  linters = expect_shape_linter()
)

lint(
  text = "expect_equal(dim(x), c(2L, 3L))",
  linters = expect_shape_linter()
)

# okay
lint(
  text = "expect_shape(x, nrow = 4L)",
  linters = expect_shape_linter()
)

lint(
  text = "expect_shape(x, dim = c(2L, 3L))",
  linters = expect_shape_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=package_development_linters]{package_development}, \link[=pkg_testthat_linters]{pkg_testthat}, \link[=readability_linters]{readability}
}
