% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_true_false_and_condition_linter.R
\name{expect_true_false_and_condition_linter}
\alias{expect_true_false_and_condition_linter}
\title{Force && conditions in expect_true(), expect_false() to be written separately}
\usage{
expect_true_false_and_condition_linter()
}
\description{
For readability of test outputs, testing only one thing per call to
\code{\link[testthat:logical-expectations]{testthat::expect_true()}} is preferable, i.e.,
\verb{expect_true(A); expect_true(B)} is better than \code{expect_true(A && B)}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=package_development_linters]{package_development}, \link[=readability_linters]{readability}
}
