% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lint.R
\name{find_default_linters}
\alias{find_default_linters}
\title{Specifying default linters}
\usage{
find_default_linters(filename)
}
\arguments{
\item{filename}{source file to be linted}
}
\description{
Lintr searches for default linters for a given file in the following order.
\enumerate{
  \item specified using the lintr.linters option.
  \item linter_file in the same directory
  \item linter_file in the project directory
  \item use the default_linters
}
}
\details{
The default linter_file name is \code{.lintr} but it can be changed with option
\code{lintr.linter_file}.
}

