% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floating_point_comparison_linter.R
\name{floating_point_comparison_linter}
\alias{floating_point_comparison_linter}
\title{Floating point comparison linter}
\usage{
floating_point_comparison_linter()
}
\description{
Avoid equality comparisons to non-integer numeric literals, which are prone to
floating point representation issues. Prefer a tolerance check such as
\code{abs(x - 3.0) < eps}.
}
\examples{
# will produce lints
lint(
  text = "x == 3.0",
  linters = floating_point_comparison_linter()
)

lint(
  text = "y == 1/10",
  linters = floating_point_comparison_linter()
)

# okay
lint(
  text = "x == 3L",
  linters = floating_point_comparison_linter()
)

lint(
  text = "x == y",
  linters = floating_point_comparison_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=robustness_linters]{robustness}
}
