% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_brace_linter.R
\name{function_brace_linter}
\alias{function_brace_linter}
\title{Require multi-line functions to use braces}
\usage{
function_brace_linter()
}
\description{
This linter catches function definitions spanning multiple lines of code
that aren't wrapped in braces
}
\seealso{
\link{linters} for a complete list of linters available in lintr. \cr
\url{https://style.tidyverse.org/syntax.html#indenting}
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
