% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_name_linters.R
\name{function_name_linter}
\alias{function_name_linter}
\title{Function name linter}
\usage{
function_name_linter(styles = c("camelCase"))
}
\arguments{
\item{styles}{A subset of
\Sexpr[stage=render, results=rd]{lintr:::regexes_rd}. A name should
match at least one of these styles.}
}
\description{
Check that function names conform to a naming style.
The default naming styles is "camelCase".
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
