% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_return_linter.R
\name{function_return_linter}
\alias{function_return_linter}
\title{Lint common mistakes/style issues cropping up from return statements}
\usage{
function_return_linter()
}
\description{
\code{return(x <- ...)} is either distracting (because \code{x} is ignored), or
confusing (because assigning to \code{x} has some side effect that is muddled
by the dual-purpose expression).
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=readability_linters]{readability}
}
