% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lint.R
\name{get_source_file}
\alias{get_source_file}
\title{Parsed sourced file from a filename}
\usage{
get_source_file(filename)
}
\arguments{
\item{filename}{the file to be parsed.}
}
\description{
This object is given as input to each linter
}

