% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_name_linters.R
\name{global_name_linter}
\alias{global_name_linter}
\title{Global variable name linter}
\usage{
global_name_linter(styles = c("SNAKE_CASE", "DOTTED.CASE"))
}
\arguments{
\item{styles}{A subset of
\Sexpr[stage=render, results=rd]{lintr:::regexes_rd}. A name should
match at least one of these styles.}
}
\description{
Check that global variable/object names conform to a naming style.
The default naming styles are either of "SNAKE_CASE" or "DOTTED.CASE".
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
