% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_null_linter.R, R/google_linters.R
\name{expect_null_linter}
\alias{expect_null_linter}
\alias{google-linters}
\title{Linters for the Google extension to the tidyverse style guide

Google uses a fork of the tidyverse style guide, pursuant to which
  its R users have written a bundle of linters, many of which
  are generally useful for writing good R code. These linters
  help encourage code consistency, readability, and best practices.}
\usage{
expect_null_linter()
}
\arguments{
\item{source_file}{A `source_file` object, as returned by [get_source_expressions()]}
}
\description{
Linters for the Google extension to the tidyverse style guide

Google uses a fork of the tidyverse style guide, pursuant to which
  its R users have written a bundle of linters, many of which
  are generally useful for writing good R code. These linters
  help encourage code consistency, readability, and best practices.
}
\section{Functions}{
\itemize{
\item \code{expect_null_linter}: Require usage of expect_null(x) over expect_equal(x, NULL) and similar
[testthat::expect_null()] exists specifically for testing for `NULL` objects.
  [testthat::expect_equal()], [testthat::expect_identical()], and
  [testthat::expect_true()] can also be used for such tests,
  but it is better to use the tailored function instead.
}}

