% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_match_braces_linter.R
\name{if_else_match_braces_linter}
\alias{if_else_match_braces_linter}
\title{Require both or neither if/else branches to use curly braces}
\usage{
if_else_match_braces_linter()
}
\description{
This linter catches \code{if}/\verb{else} clauses where the \code{if} branch is wrapped
in \code{{...}} but the \verb{else} branch is not, or vice versa, i.e., it ensures
that either both branches use \code{{...}} or neither does.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=readability_linters]{readability}, \link[=style_linters]{style}
}
