% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_switch_linter.R
\name{if_switch_linter}
\alias{if_switch_linter}
\title{Require usage of switch() over repeated if/else blocks}
\usage{
if_switch_linter()
}
\description{
\code{\link[=switch]{switch()}} statements in R are used to delegate behavior based
on the value of some input scalar string, e.g.
\code{switch(x, a = 1, b = 3, c = 7, d = 8)} will be one of
\code{1}, \code{3}, \code{7}, or \code{8}, depending on the value of \code{x}.
}
\details{
This can also be accomplished by repeated \code{if}/\verb{else} statements like
so: \code{if (x == "a") 1 else if (x == "b") 2 else if (x == "c") 7 else 8}
(implicitly, the last \verb{else} assumes x only takes 4 possible values),
but this is more cluttered and slower (note that \code{switch()} takes the same
time to evaluate regardless of the value of \code{x}, and is faster even
when \code{x} takes the first value (here \code{a}), and that the \code{if}/\verb{else}
approach is roughly linear in the number of conditions that need to
be evaluated, here up to 3 times).
}
\examples{
# will produce lints
lint(
  text = "if (x == 'a') 1 else if (x == 'b') 2 else 3",
  linters = if_switch_linter()
)

# okay
lint(
  text = "if (x == 'a') 1 else if (x == 'b' & y == 2) 2 else 3",
  linters = if_switch_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}
}
