% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_else_return_linter.R
\name{implicit_else_return_linter}
\alias{implicit_else_return_linter}
\title{Require an explicit call to else in a terminal if statement}
\usage{
implicit_else_return_linter()
}
\description{
When an \code{if} statement in R is missing \verb{else}, the alternative is implicitly
set to \code{NULL}. When the \code{if} statement comes at the end of a function
definition, then, there is an implicit return of \code{NULL}, which violates
\href{https://goto.google.com/rstyle#use-explicit-returns.}{go/rstyle#use-explicit-returns.}
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
