% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_linter.R
\name{in_linter}
\alias{in_linter}
\title{Chained equality check linter}
\usage{
in_linter()
}
\description{
Report the use of chained equality checks where \code{\%in\%} would be more
appropriate
}
\examples{
# lints
lint(
  text = "x == 'a' | x == 'b'",
  linters = in_linter()
)

# This only makes sense in the case x if of length 1
lint(
  text = "x == 'a' || x == 'b'",
  linters = in_linter()
)

# okay
lint(
  text = "x \%in\% c('a', 'b')",
  linters = in_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}
}
