% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_comparison_linter.R
\name{inner_comparison_linter}
\alias{inner_comparison_linter}
\title{Require == to be used outside of sapply() when comparing to a constant}
\usage{
inner_comparison_linter()
}
\description{
\code{sapply(x, function(xi) foo(xi) == 2)} is the same as \code{sapply(x, foo) == 2},
but misses the opportunity to vectorize the call to \code{==}.
}
\examples{
# will produce lints
lint(
  text = "sapply(x, function(xi) xi == 2)",
  linters = inner_comparison_linter()
)

lint(
  text = "sapply(x, function(xi) sum(xi) > 0)",
  linters = inner_comparison_linter()
)

# okay
lint(
  text = "sapply(x, sum) > 0",
  linters = inner_comparison_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=efficiency_linters]{efficiency}
}
