% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lint_level.R
\name{is_lint_level}
\alias{is_lint_level}
\title{Is this an expression- or a file-level source object?}
\usage{
is_lint_level(
  source_expression,
  level = c("expression", "file"),
  require_xml = FALSE
)
}
\arguments{
\item{source_expression}{A parsed expression object, i.e., an element
of the object returned by \code{\link[=get_source_expressions]{get_source_expressions()}}.}

\item{level}{Which level of expression is being tested? \code{"expression"}
means an individual expression, while \code{"file"} means the full
source tree for this file.}

\item{require_xml}{If \code{TRUE}, the XML equivalent of the parse tree
is required. Use this whenever writing a linter based on logic
expressed via Xpath.}
}
\description{
Helper for determining whether the current \code{source_expression} contains
the full file tree, or is just a single expression.
}
