% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric_linter.R
\name{is_numeric_linter}
\alias{is_numeric_linter}
\title{Redirect \code{is.numeric(x) || is.integer(x)} to just use \code{is.numeric(x)}}
\usage{
is_numeric_linter()
}
\description{
\code{\link[=is.numeric]{is.numeric()}} returns \code{TRUE} when \code{typeof(x)} is \code{double} or \code{integer} --
testing \code{is.numeric(x) || is.integer(x)} is thus redundant.
}
