% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_test_linter.R
\name{length_test_linter}
\alias{length_test_linter}
\title{Check for a common mistake where length is applied in the wrong place}
\usage{
length_test_linter()
}
\description{
Usage like \code{length(x == 0)} is a mistake. If checking the length of \code{x},
\code{length(x) == 0} was intended. The length of the logical vector \code{x == 0}
will be the same as the length of \code{x}, so it would be more direct to use
\code{length(x)} if that value was really intended. Even in more complicated
cases like \code{length(x == y)} where \code{x} and \code{y} may have different lengths
and this expression relies on recycling rules, it is far preferable to
make the intent clearer, e.g. \code{max(length(x), length(y))}.
}
\examples{
# will produce lints
lint(
  text = "length(x == 0)",
  linters = length_test_linter()
)

# okay
lint(
  text = "length(x) > 0",
  linters = length_test_linter()
)
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=robustness_linters]{robustness}
}
