% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths_linter.R
\name{lengths_linter}
\alias{lengths_linter}
\title{Require usage of \code{lengths()} where possible}
\usage{
lengths_linter()
}
\description{
\code{\link[=lengths]{lengths()}} is a function that was added to base R in version 3.2.0 to
get the length of each element of a list. It is equivalent to
\code{sapply(x, length)}, but faster and more readable.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}
}
