% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_call_linter.R
\name{library_call_linter}
\alias{library_call_linter}
\title{Library call linter}
\usage{
library_call_linter()
}
\description{
Force library calls to all be at the top of the script.
}
\examples{
# will produce lints
lint(
  text = "
    library(dplyr)
    print('test')
    library(tidyr)
  ",
  linters = library_call_linter()
)

lint(
  text = "
    library(dplyr)
    print('test')
    library(tidyr)
    library(purrr)
  ",
  linters = library_call_linter()
)

# okay
lint(
  text = "
    library(dplyr)
    print('test')
  ",
  linters = library_call_linter()
)

lint(
  text = "
    # comment
    library(dplyr)
  ",
  linters = library_call_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=style_linters]{style}
}
