% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_file}
\alias{lint}
\alias{lint_file}
\title{Lint a given file}
\usage{
lint(filename, linters = NULL, cache = FALSE, exclusions = NULL, ...)
}
\arguments{
\item{filename}{the given filename to lint.}

\item{linters}{a list of linter functions to apply see \code{\link{linters}}
for a full list of default and available linters.}

\item{cache}{toggle caching of lint results}

\item{exclusions}{additional lines or files to exclude.}

\item{...}{additional arguments passed to \code{\link{parse_exclusions}}.}
}
\description{
Apply one or more linters to a file and return a list of lints found.
}
\section{Exclusions}{

Exclusions can be specified in three different ways.
\enumerate{
\item{single line in the source file. default: \code{# nolint}}
\item{line range in the source file. default: \code{# nolint start}, \code{# nolint end}}
\item{exclusions parameter, a named list of the files and lines to exclude, or just the filenames
if you want to exclude the entire file.}
}
}

