% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr-deprecated.R
\name{lintr-deprecated}
\alias{lintr-deprecated}
\alias{closed_curly_linter}
\alias{open_curly_linter}
\alias{paren_brace_linter}
\alias{semicolon_terminator_linter}
\title{Deprecated functions in lintr}
\usage{
closed_curly_linter(allow_single_line = FALSE)

open_curly_linter(allow_single_line = FALSE)

paren_brace_linter()

semicolon_terminator_linter(semicolon = c("compound", "trailing"))
}
\arguments{
\item{allow_single_line}{if \code{TRUE}, allow an open and closed curly pair on the same line.}

\item{semicolon}{A character vector defining which semicolons to report:
\describe{
\item{compound}{Semicolons that separate two statements on the same line.}
\item{trailing}{Semicolons following the last statement on the line.}
}}
}
\description{
These functions have been deprecated from lintr.
\itemize{
\item \code{open_curly_linter()} and \code{closed_curly_linter()} check that open and closed curly braces
are on their own line unless they follow an else, a comma, or a closing bracket.
Deprecated in favour of \code{brace_linter()}.
\item \code{paren_brace_linter()} checks that there is a space between right parentheses and an opening
curly brace. E.g., \code{function(){}} doesn't have a space, while \code{function() {}} does.
Deprecated in favour of \code{brace_linter()}.
\item \code{semicolon_terminator_linter()} checks that no semicolons terminate expressions.
Deprecated in favour of \code{semicolon_linter()}.
}
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\keyword{internal}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=deprecated_linters]{deprecated}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
