% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lint.R
\name{lintr}
\alias{lintr}
\title{Lintr}
\description{
Static code analysis to find errors in style, syntax and semantics.
}
\seealso{
\code{\link{lint}}, \code{\link{lint_package}}, \code{\link{linters}}
}

