% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{make_linter_message}
\alias{make_linter_message}
\title{Construct a Linter Message}
\usage{
make_linter_message(header, content, lines)
}
\arguments{
\item{header}{A header message describing the linter.}

\item{content}{The content of the file that was linted.}

\item{lines}{The line numbers from \code{content} that contain lint.}
}
\description{
Pretty-prints a linter message. Primarily used as a helper
for constructing linter messages with \code{\link{linter}}.
}

