% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_argument_linter.R
\name{missing_argument_linter}
\alias{missing_argument_linter}
\title{Missing argument linter}
\usage{
missing_argument_linter(
  except = c("alist", "quote", "rowwiseDT", "switch"),
  allow_trailing = FALSE
)
}
\arguments{
\item{except}{a character vector of function names as exceptions.}

\item{allow_trailing}{always allow trailing empty arguments?}
}
\description{
Check for missing arguments in function calls (e.g. \code{stats::median(1:10, )}).
}
\examples{
# will produce lints
lint(
  text = 'tibble(x = "a", )',
  linters = missing_argument_linter()
)

# okay
lint(
  text = 'tibble(x = "a")',
  linters = missing_argument_linter()
)

lint(
  text = 'tibble(x = "a", )',
  linters = missing_argument_linter(except = "tibble")
)

lint(
  text = 'tibble(x = "a", )',
  linters = missing_argument_linter(allow_trailing = TRUE)
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=configurable_linters]{configurable}, \link[=correctness_linters]{correctness}
}
