% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_pipe_linter.R
\name{nested_pipe_linter}
\alias{nested_pipe_linter}
\title{Block usage of pipes nested inside other calls}
\usage{
nested_pipe_linter()
}
\description{
Nesting pipes harms readability; extract sub-steps to separate variables,
append further pipeline steps, or otherwise refactor such usage away.
}
\examples{
# will produce lints
lint(
  text = "df1 \%>\% inner_join(df2 \%>\% select(a, b))",
  linters = nested_pipe_linter()
)

# okay
lint(
  text = "tryCatch(x \%>\% filter(grp == 'a'), error = identity)",
  linters = nested_pipe_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
