% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_tab_linter.R
\name{no_tab_linter}
\alias{no_tab_linter}
\title{No tab linter}
\usage{
no_tab_linter()
}
\description{
Check that only spaces are used for indentation, not tabs.
}
\examples{
# will produce lints
lint(
  text = "\tx",
  linters = no_tab_linter()
)

# okay
lint(
  text = "  x",
  linters = no_tab_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
