% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonportable_path_linter.R
\name{nonportable_path_linter}
\alias{nonportable_path_linter}
\title{Non-portable path linter}
\usage{
nonportable_path_linter(lax = TRUE)
}
\arguments{
\item{lax}{Less stringent linting, leading to fewer false positives.
If \code{TRUE}, only lint path strings, which
\itemize{
\item contain at least two path elements, with one having at least two characters and
\item contain only alphanumeric chars (including UTF-8), spaces, and win32-allowed punctuation
}}
}
\description{
Check that \code{\link[=file.path]{file.path()}} is used to construct safe and portable paths.
}
\examples{
# will produce lints
lint(
  text = "'abcdefg/hijklmnop/qrst/uv/wxyz'",
  linters = nonportable_path_linter()
)

# okay
lint(
  text = "file.path('abcdefg', 'hijklmnop', 'qrst', 'uv', 'wxyz')",
  linters = nonportable_path_linter()
)

}
\seealso{
\itemize{
\item \link{linters} for a complete list of linters available in lintr.
\item \code{\link[=absolute_path_linter]{absolute_path_linter()}}
}
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=configurable_linters]{configurable}, \link[=robustness_linters]{robustness}
}
