% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_name_linters.R
\name{object_name_linter}
\alias{object_name_linter}
\title{Object name linter}
\usage{
object_name_linter(styles = c("snake_case", "symbols"))
}
\arguments{
\item{styles}{A subset of
\Sexpr[stage=render, results=rd]{lintr:::regexes_rd}. A name should
match at least one of these styles.}
}
\description{
Check that object names conform to a naming style.
The default naming styles are "snake_case" and "symbols".
}
