% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_curly_linter.R
\name{open_curly_linter}
\alias{open_curly_linter}
\title{Open curly linter}
\usage{
open_curly_linter(allow_single_line = FALSE)
}
\arguments{
\item{allow_single_line}{if \code{TRUE}, allow an open and closed curly pair on the same line.}
}
\description{
Check that opening curly braces are never on their own line and are always followed by a newline.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
