% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer_negation_linter.R
\name{outer_negation_linter}
\alias{outer_negation_linter}
\title{Require usage of !any(.) over all(!.), !all(.) over any(!.)}
\usage{
outer_negation_linter()
}
\description{
\code{any(!x)} is logically equivalent to \code{!any(x)}; ditto for the equivalence of
\code{all(!x)} and \code{!any(x)}. Negating after aggregation only requires inverting
one logical value, and should usually be more readable.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}
}
