% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paren_body_linter.R
\name{paren_body_linter}
\alias{paren_body_linter}
\title{Parenthesis before body linter}
\usage{
paren_body_linter()
}
\description{
Check that there is a space between right parenthesis and a body expression.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
