% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paren_brace_linter.R
\name{paren_brace_linter}
\alias{paren_brace_linter}
\title{Parentheses before brace linter}
\usage{
paren_brace_linter()
}
\description{
Check that there is a space between right parentheses and an opening curly brace.
For example, \code{function(){}} doesn't have a space, while \code{function() {}} does.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=deprecated_linters]{deprecated}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
