% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_sep_linter.R
\name{paste_sep_linter}
\alias{paste_sep_linter}
\title{Block usage of paste() with sep=""}
\usage{
paste_sep_linter()
}
\description{
\code{paste0} is a faster, more concise alternative to using \code{paste(sep = "")}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}
}
