% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_to_string_linter.R
\name{paste_to_string_linter}
\alias{paste_to_string_linter}
\title{Block usage of paste() with collapse=", "}
\usage{
paste_to_string_linter()
}
\description{
\code{\link[=toString]{toString()}} is a more concise and expressive wrapper for aggregating string
vectors with \code{paste(., collapse = ", ")}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}
}
