% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_call_linter.R
\name{pipe_call_linter}
\alias{pipe_call_linter}
\title{Pipe call linter}
\usage{
pipe_call_linter()
}
\description{
Force explicit calls in magrittr pipes, e.g.,
\code{1:3 \%>\% sum()} instead of \code{1:3 \%>\% sum}.
}
