% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_consistency_linter.R
\name{pipe_consistency_linter}
\alias{pipe_consistency_linter}
\title{Pipe consistency linter}
\usage{
pipe_consistency_linter(pipe = c("auto", "\%>\%", "|>"))
}
\arguments{
\item{pipe}{Which pipe operator to use (either \code{"\%>\%"} or \code{"|>"}). By default
(`"auto"``), the linter has no preference but will check that the same pipe
operator is used throughout the file.}
}
\description{
Check that pipe operators are consistently used throughout a file. Optionally
specify which pipe operator to use.
}
\examples{
# will produce lints
lint(
  text = "1:3 |> mean() \%>\% as.character()",
  linters = pipe_consistency_linter()
)

lint(
  text = "1:3 \%>\% mean() \%>\% as.character()",
  linters = pipe_consistency_linter("|>")
)

# okay
lint(
  text = "1:3 \%>\% mean() \%>\% as.character()",
  linters = pipe_consistency_linter()
)

lint(
 text = "1:3 |> mean() |> as.character()",
 linters = pipe_consistency_linter()
)
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
