% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{portability_linters}
\alias{portability_linters}
\title{Portability linters}
\description{
Linters that check for code portability issues, i.e., code patterns that might fail, or work differently,
on different platforms.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'portability':
\itemize{
\item{\code{\link{absolute_path_linter}}}
\item{\code{\link{all_equal_linter}}}
\item{\code{\link{backport_linter}}}
\item{\code{\link{download_file_linter}}}
\item{\code{\link{nonportable_path_linter}}}
\item{\code{\link{paste_linter}}}
\item{\code{\link{pipe_consistency_linter}}}
\item{\code{\link{strings_as_factors_linter}}}
\item{\code{\link{system_file_linter}}}
}
}
