% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{readability_linters}
\alias{readability_linters}
\title{Readability linters}
\description{
Linters highlighting readability issues, such as missing whitespace.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'readability':
\itemize{
\item{\code{\link{boolean_arithmetic_linter}}}
\item{\code{\link{brace_linter}}}
\item{\code{\link{coalesce_linter}}}
\item{\code{\link{commas_linter}}}
\item{\code{\link{commented_code_linter}}}
\item{\code{\link{comparison_negation_linter}}}
\item{\code{\link{conjunct_test_linter}}}
\item{\code{\link{consecutive_assertion_linter}}}
\item{\code{\link{consecutive_mutate_linter}}}
\item{\code{\link{cyclocomp_linter}}}
\item{\code{\link{empty_assignment_linter}}}
\item{\code{\link{expect_length_linter}}}
\item{\code{\link{expect_named_linter}}}
\item{\code{\link{expect_not_linter}}}
\item{\code{\link{expect_shape_linter}}}
\item{\code{\link{expect_true_false_linter}}}
\item{\code{\link{fixed_regex_linter}}}
\item{\code{\link{for_loop_index_linter}}}
\item{\code{\link{function_left_parentheses_linter}}}
\item{\code{\link{function_return_linter}}}
\item{\code{\link{if_not_else_linter}}}
\item{\code{\link{if_switch_linter}}}
\item{\code{\link{implicit_assignment_linter}}}
\item{\code{\link{indentation_linter}}}
\item{\code{\link{infix_spaces_linter}}}
\item{\code{\link{inner_combine_linter}}}
\item{\code{\link{is_numeric_linter}}}
\item{\code{\link{keyword_quote_linter}}}
\item{\code{\link{length_levels_linter}}}
\item{\code{\link{lengths_linter}}}
\item{\code{\link{library_call_linter}}}
\item{\code{\link{line_length_linter}}}
\item{\code{\link{list2df_linter}}}
\item{\code{\link{matrix_apply_linter}}}
\item{\code{\link{nested_ifelse_linter}}}
\item{\code{\link{nested_pipe_linter}}}
\item{\code{\link{numeric_leading_zero_linter}}}
\item{\code{\link{object_length_linter}}}
\item{\code{\link{object_overwrite_linter}}}
\item{\code{\link{object_usage_linter}}}
\item{\code{\link{one_call_pipe_linter}}}
\item{\code{\link{outer_negation_linter}}}
\item{\code{\link{paren_body_linter}}}
\item{\code{\link{pipe_call_linter}}}
\item{\code{\link{pipe_consistency_linter}}}
\item{\code{\link{pipe_continuation_linter}}}
\item{\code{\link{quotes_linter}}}
\item{\code{\link{redundant_equals_linter}}}
\item{\code{\link{rep_len_linter}}}
\item{\code{\link{repeat_linter}}}
\item{\code{\link{sample_int_linter}}}
\item{\code{\link{scalar_in_linter}}}
\item{\code{\link{semicolon_linter}}}
\item{\code{\link{sort_linter}}}
\item{\code{\link{spaces_inside_linter}}}
\item{\code{\link{spaces_left_parentheses_linter}}}
\item{\code{\link{stopifnot_all_linter}}}
\item{\code{\link{string_boundary_linter}}}
\item{\code{\link{system_file_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
\item{\code{\link{unnecessary_concatenation_linter}}}
\item{\code{\link{unnecessary_lambda_linter}}}
\item{\code{\link{unnecessary_nesting_linter}}}
\item{\code{\link{unnecessary_placeholder_linter}}}
\item{\code{\link{unreachable_code_linter}}}
\item{\code{\link{which_grepl_linter}}}
\item{\code{\link{yoda_test_linter}}}
}
}
