% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant_equals_linter.R
\name{redundant_equals_linter}
\alias{redundant_equals_linter}
\title{Block usage of ==, != on logical vectors}
\usage{
redundant_equals_linter()
}
\description{
Testing \code{x == TRUE} is redundant if \code{x} is a logical vector. Wherever this
is used to improve readability, the solution should instead be to improve
the naming of the object to better indicate that its contents are logical.
}
