% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{regex_linters}
\alias{regex_linters}
\title{Regular expression linters}
\description{
Linters that examine the usage of regular expressions and functions executing them in user code.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'regex':
\itemize{
\item{\code{\link{fixed_regex_linter}}}
\item{\code{\link{regex_subset_linter}}}
\item{\code{\link{string_boundary_linter}}}
\item{\code{\link{which_grepl_linter}}}
}
}
