% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_len_linter.R
\name{rep_len_linter}
\alias{rep_len_linter}
\title{Require usage of rep_len(x, n) over rep(x, length.out = n)}
\usage{
rep_len_linter()
}
\description{
\code{rep(x, length.out = n)} calls \code{rep_len(x, n)} "under the hood". The latter
is thus more direct and equally readable. See also \code{\link[=SampleIntLinter]{SampleIntLinter()}}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
