% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assignment_linter.R
\name{return_assignment_linter}
\alias{return_assignment_linter}
\title{Block assignments inside function returns}
\usage{
return_assignment_linter()
}
\description{
\code{return(x <- ...)} is either distracting (because \code{x} is ignored), or
confusing (because assigning to \code{x} has some side effect that is muddled
by the dual-purpose expression).
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=readability_linters]{readability}
}
