% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_linter.R
\name{return_linter}
\alias{return_linter}
\title{Return linter}
\usage{
return_linter(
  use_implicit_returns = TRUE,
  additional_allowed_func = NULL,
  additional_side_effect_func = NULL,
  use_runit = FALSE
)
}
\arguments{
\item{use_implicit_returns}{Whether to use implicit or explicit returns}

\item{additional_allowed_func}{Names of additional functions that are
accepted as return if \code{!use_implicit_returns}}

\item{additional_side_effect_func}{Names of additional functions that are
not checked for an explicit retun if \code{!use_implicit_returns}}

\item{use_runit}{Whether to ignore Runit like functions or not
if \code{!use_implicit_returns}}
}
\description{
This linter checks for explicit \code{\link[=return]{return()}} at the end of a function
}
\examples{
# will produce lints
code <- "function(x) {\n  return(x + 1)\n}"
writeLines(code)
lint(
  text = code,
  linters = return_linter()
)

code <- "function(x) {\n  x + 1\n}"
writeLines(code)
lint(
  text = code,
  linters = return_linter(use_implicit_returns = FALSE)
)

# okay
code <- "function(x) {\n  x + 1\n}"
writeLines(code)
lint(
  text = code,
  linters = return_linter()
)

code <- "function(x) {\n  return(x + 1)\n}"
writeLines(code)
lint(
  text = code,
  linters = return_linter(use_implicit_returns = FALSE)
)


}
\seealso{
\itemize{
\item \link{linters} for a complete list of linters available in lintr.
\item \url{https://style.tidyverse.org/functions.html?q=return#return}
}
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
