% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{sarif_output}
\alias{sarif_output}
\title{SARIF Report for lint results}
\usage{
sarif_output(lints, filename = "lintr_results.sarif")
}
\arguments{
\item{lints}{the linting results.}

\item{filename}{the name of the output report}
}
\description{
Generate a report of the linting results using the \href{https://sarifweb.azurewebsites.net/}{SARIF} format.
}
