% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicolon_terminator_linter.R
\name{semicolon_terminator_linter}
\alias{semicolon_terminator_linter}
\title{Semicolon terminator linter}
\usage{
semicolon_terminator_linter(semicolon = c("compound", "trailing"))
}
\arguments{
\item{semicolon}{A character vector defining which semicolons to report:\describe{
  \item{compound}{Semicolons that separate two statements on the same line.}
  \item{trailing}{Semicolons following the last statement on the line.}
}}
}
\description{
Check that no semicolons terminate expressions.
}
