% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_quotes_linter.R
\name{single_quotes_linter}
\alias{single_quotes_linter}
\title{Single quotes linter}
\usage{
single_quotes_linter()
}
\description{
Check that only double quotes are used to delimit string constants.
}
