% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation_linter.R
\name{stack_to_segments}
\alias{stack_to_segments}
\title{Convert a stack of bookends into independent segments}
\usage{
stack_to_segments(x)
}
\description{
Given a stack, provided as a vector of positive and negative bookend
identifiers, convert the stack into a vector of identifiers highest on the
stack at each position.
}
\examples{
# Stack:                                 |--- 7 ---|
#                                |----------- 4 -------|
#            |--- 1 ---|     |--------------- 9 ------------| |----- 3 -----|
x <- c(0, 0,  1,  0, -1,  0,  9,  4,  0,  7,  0, -7, -4,  -9,  3,  0,  0, -3)
stack_to_segments(x)
#> [1] 0  0   1   1   1   0   9   4   4   7   7   7   4    9   3   3   3   3
# Segments:  |--- 1 ---|     |9| |- 4 -| |--- 7 ---| |4|  |9| |----- 3 -----|

x <- c(0, 1, 2, 0, 0, -2, -1,  0)
stack_to_segments(x)
#> [1] 0 1 2 2 2 2 1 0


}
