% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_paste_linter.R
\name{stop_paste_linter}
\alias{stop_paste_linter}
\title{Block usage of paste() and paste0() with messaging functions using ...}
\usage{
stop_paste_linter()
}
\description{
\code{stop(paste0(...))} is strictly redundant -- \code{stop(...)} is equivalent.
\code{stop(...)} is also preferable to \code{stop(paste(...))}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}
}
