% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnot_conjunct_linter.R
\name{stopifnot_conjunct_linter}
\alias{stopifnot_conjunct_linter}
\title{Force && conditions in stopifnot() to be written separately}
\usage{
stopifnot_conjunct_linter()
}
\description{
\code{stopifnot()} accepts any number of tests, so sequences like
\code{stopifnot(x && y)} are better written as \code{stopifnot(x, y)} because
the error message in the latter case is more useful.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
No tags are given.
}
