% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{style_linters}
\alias{style_linters}
\title{Style linters}
\description{
Linters highlighting code style issues.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'style':
\itemize{
\item{\code{\link{assignment_linter}}}
\item{\code{\link{brace_linter}}}
\item{\code{\link{commas_linter}}}
\item{\code{\link{commented_code_linter}}}
\item{\code{\link{complex_conditional_linter}}}
\item{\code{\link{condition_call_linter}}}
\item{\code{\link{consecutive_assertion_linter}}}
\item{\code{\link{cyclocomp_linter}}}
\item{\code{\link{function_argument_linter}}}
\item{\code{\link{function_left_parentheses_linter}}}
\item{\code{\link{implicit_assignment_linter}}}
\item{\code{\link{implicit_integer_linter}}}
\item{\code{\link{indentation_linter}}}
\item{\code{\link{infix_spaces_linter}}}
\item{\code{\link{keyword_quote_linter}}}
\item{\code{\link{library_call_linter}}}
\item{\code{\link{line_length_linter}}}
\item{\code{\link{numeric_leading_zero_linter}}}
\item{\code{\link{object_length_linter}}}
\item{\code{\link{object_name_linter}}}
\item{\code{\link{object_usage_linter}}}
\item{\code{\link{one_call_pipe_linter}}}
\item{\code{\link{package_hooks_linter}}}
\item{\code{\link{paren_body_linter}}}
\item{\code{\link{pipe_call_linter}}}
\item{\code{\link{pipe_consistency_linter}}}
\item{\code{\link{pipe_continuation_linter}}}
\item{\code{\link{quotes_linter}}}
\item{\code{\link{repeat_linter}}}
\item{\code{\link{return_linter}}}
\item{\code{\link{semicolon_linter}}}
\item{\code{\link{spaces_inside_linter}}}
\item{\code{\link{spaces_left_parentheses_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
\item{\code{\link{todo_comment_linter}}}
\item{\code{\link{trailing_blank_lines_linter}}}
\item{\code{\link{trailing_whitespace_linter}}}
\item{\code{\link{undesirable_function_linter}}}
\item{\code{\link{undesirable_operator_linter}}}
\item{\code{\link{unnecessary_concatenation_linter}}}
\item{\code{\link{whitespace_linter}}}
}
}
