% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_close_linter.R
\name{terminal_close_linter}
\alias{terminal_close_linter}
\title{Prohibit close() from terminating a function definition}
\usage{
terminal_close_linter()
}
\description{
Functions that end in \code{close(x)} are almost always better written by using
\code{on.exit(close(x))} close to where \code{x} is defined and/or opened.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_bractices_linters]{best_bractices}, \link[=robustness_linters]{robustness}
}
