% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{tidy_design_linters}
\alias{tidy_design_linters}
\title{Tidyverse design linters}
\description{
Linters based on guidelines described in the 'Tidy design principles' book.
}
\seealso{
\itemize{
\item \link{linters} for a complete list of linters available in lintr.
\item \url{https://design.tidyverse.org/}
}
}
\section{Linters}{
The following linters are tagged with 'tidy_design':
\itemize{
\item{\code{\link{condition_call_linter}}}
}
}
