% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailing_assignment_linter.R
\name{trailing_assignment_linter}
\alias{trailing_assignment_linter}
\title{Trailing assignment linter}
\usage{
trailing_assignment_linter(allow_piping = TRUE, allow_comments = FALSE)
}
\arguments{
\item{allow_piping}{Suppress lints when assignment begins/ends a pipe}

\item{allow_comments}{Suppress lints for commented out lines}
}
\description{
Check that there are no assignment operators at the end of source lines.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=style_linters]{style}
}
