% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnecessary_placeholder_linter.R
\name{unnecessary_placeholder_linter}
\alias{unnecessary_placeholder_linter}
\title{Block usage of pipeline placeholders if unnecessary}
\usage{
unnecessary_placeholder_linter()
}
\description{
The argument placeholder \code{.} in magrittr pipelines is unnecessary if
passed as the first positional argument; using it can cause confusion
and impacts readability.
}
