% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_name_linters.R
\name{variable_name_linter}
\alias{variable_name_linter}
\title{Variable name linter}
\usage{
variable_name_linter(styles = c("snake_case", "dotted.case"))
}
\arguments{
\item{styles}{A subset of
\Sexpr[stage=render, results=rd]{lintr:::regexes_rd}. A name should
match at least one of these styles.}
}
\description{
Check that non-global and non-function object names conform to a naming style.
The default naming styles are either of "snake_case" or "dotted.case".
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
