% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_linter.R
\name{which_linter}
\alias{which_linter}
\title{Find suboptimal usages involving \code{which()}}
\usage{
which_linter()
}
\description{
\itemize{
\item \code{which.min(x)} and \code{which.max(x)} are optimized functions that directly return
the index of the minimum or maximum value in a vector. They are more
efficient and cleaner than using \code{which(min(x))} or \code{which(max(x))}.
\item \code{which(grepl(pattern, x))} is the same as \code{grep(pattern, x)}, but harder
to read and requires two passes over the vector.
}
}
\examples{
# will produce lints
lint(
  text = "which(x == max(x))",
  linters = unnecessary_which_linter()
)

lint(
  text = "which(x == min(x))",
  linters = unnecessary_which_linter()
)

lint(
  text = "which(grepl('^a', x))",
  linters = which_linter()
)

# okay
lint(
  text = "which.max(x))",
  linters = unnecessary_which_linter()
)

lint(
  text = "which.min(x))",
  linters = unnecessary_which_linter()
)

lint(
  text = "which(grepl('^a', x) | grepl('^b', x))",
  linters = which_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}, \link[=regex_linters]{regex}
}
