% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitespace_linter.R
\name{whitespace_linter}
\alias{whitespace_linter}
\title{Whitespace linter}
\usage{
whitespace_linter()
}
\description{
Check that the correct character is used for indentation.
}
\details{
Currently, only supports linting in the presence of tabs.

Much ink has been spilled on this topic, and we encourage you to check
out references for more information.
}
\examples{
# will produce lints
lint(
  text = "\tx",
  linters = whitespace_linter()
)

# okay
lint(
  text = "  x",
  linters = whitespace_linter()
)

}
\references{
\itemize{
\item https://www.jwz.org/doc/tabs-vs-spaces.html
\item https://blog.codinghorror.com/death-to-the-space-infidels/
}
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
