% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{with_defaults}
\alias{with_defaults}
\title{Modify the list of default linters}
\usage{
with_defaults(..., default = default_linters)
}
\arguments{
\item{...}{named arguments of linters to change.  If the named linter already
exists it is replaced by the new linter, if it does not exist it is added.
If the value is \code{NULL} the linter is removed.}

\item{default}{default linters to change}
}
\description{
Modify the list of default linters
}
\examples{
# change the default line length cutoff
with_defaults(line_length_linter = line_length_linter(120))

# you can also omit the argument name if you are just using different
#   arguments.
with_defaults(line_length_linter(120))

# enforce camelCase rather than snake_case
with_defaults(camel_case_linter = NULL,
              snake_case_linter)
}

