% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_tags}
\alias{with_tags}
\title{Create a tag-based linter configuration}
\usage{
with_tags(..., tags = "default", packages = "lintr")
}
\arguments{
\item{...}{Arguments of elements to change. If unnamed, the argument is automatically named.
If the named argument already exists in the list of linters, it is replaced by the new element.
If it does not exist, it is added. If the value is \code{NULL}, the linter is removed.}

\item{tags}{Optional character vector of tags to search. Only linters with at least one matching tag will be
returned. If \code{tags} is \code{NULL}, all linters will be returned.}

\item{packages}{A character vector of packages to search for linters.}
}
\value{
A modified list of linters.
}
\description{
Make a new list based on all linters tagged with \code{tags} from provided by \code{packages}.
The result of this function is meant to be passed to the \code{linters} argument of \code{lint()}, or put in your
configuration file.
}
\examples{
# Default invocations of `with_defaults()` and `with_tags()` are the same:
all.equal(with_defaults(), with_tags())

# Get all linters useful for package development
with_tags(tags = "package_development")

# Get all linters provided by lintr
with_tags(tags = NULL)

# Get all linters tagged as "default" from lintr and mypkg
\dontrun{with_tags(packages = c("lintr", "mypkg"))}
}
\seealso{
\link{with_defaults} for basing off lintrs set of default linters.

\link{available_linters} to get a data frame of available linters.

\link{linters} for a complete list of linters available in lintr.
}
