% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_utils.R
\name{xml_nodes_to_lints}
\alias{xml_nodes_to_lints}
\title{Convert an XML node or nodeset into a Lint}
\usage{
xml_nodes_to_lints(
  xml,
  source_expression,
  lint_message,
  type = c("style", "warning", "error"),
  offset = 0L
)
}
\arguments{
\item{xml}{An \code{xml_node} object (to generate one \code{Lint}) or an
\code{xml_nodeset} object (to generate several \code{Lint}s), e.g. as returned by
\code{\link[xml2:xml_find_all]{xml2::xml_find_all()}} or \code{\link[xml2:xml_find_all]{xml2::xml_find_first()}}.}

\item{source_expression}{A source expression object, e.g. as
returned typically by \code{\link[=lint]{lint()}}, or more generally
by \code{\link[=get_source_expressions]{get_source_expressions()}}.}

\item{lint_message}{The message to be included as the \code{message}
to the \code{Lint} object. If \code{lint_message} is a \code{function},
this function is first applied to \code{xml} (so it should be a
function taking an \code{xml_node} as input and must produce a
length-1 character as output).}

\item{type}{type of lint.}

\item{offset}{Integer, default 0. The amount by which to offset the
\code{column_number} relative to \code{col1} (useful for getting the RStudio
source markers to land in a particular place for convenient editing,
for example). Adjustments will also be made to ensure
\verb{ranges[1L] <= column_number <= ranges[2L]}.}
}
\value{
For \code{xml_node}s, a \code{lint}. For \code{xml_nodeset}s, \code{lints} (a list of \code{lint}s).
}
\description{
Convenience function for converting nodes matched by XPath-based
linter logic into a \code{\link[=Lint]{Lint()}} object to return.
}
