% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_linters.R
\name{absolute_path_linter}
\alias{absolute_path_linter}
\title{Absolute path linter}
\usage{
absolute_path_linter(lax = TRUE)
}
\arguments{
\item{lax}{Less stringent linting, leading to fewer false positives.}
}
\description{
Check that no absolute paths are used (e.g. "/var", "C:\\System", "~/docs").
}
