% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_linters.R
\name{absolute_path_linter}
\alias{absolute_path_linter}
\title{Absolute path linter}
\usage{
absolute_path_linter(lax = TRUE)
}
\arguments{
\item{lax}{Less stringent linting, leading to fewer false positives.
If \code{TRUE}, only lint path strings, which
\itemize{
\item contain at least two path elements, with one having at least two characters and
\item contain only alphanumeric chars (including UTF-8), spaces, and win32-allowed punctuation
}}
}
\description{
Check that no absolute paths are used (e.g. "/var", "C:\\System", "~/docs").
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=configurable_linters]{configurable}, \link[=robustness_linters]{robustness}
}
