% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_linter.R
\name{assignment_linter}
\alias{assignment_linter}
\title{Assignment linter}
\usage{
assignment_linter(block_double_assign = FALSE)
}
\arguments{
\item{block_double_assign}{Logical, \code{FALSE} by default. If \code{TRUE}, usage
of \code{\link[base:assignOps]{<<-}} and \code{\link[base:assignOps]{->>}} are also blocked.}
}
\description{
Check that \verb{<-} is always used for assignment.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
