% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_linter.R
\name{assignment_linter}
\alias{assignment_linter}
\title{Assignment linter}
\usage{
assignment_linter(allow_cascading_assign = TRUE, allow_right_assign = FALSE)
}
\arguments{
\item{allow_cascading_assign}{Logical, default \code{TRUE}.
If \code{FALSE}, \code{\link[base:assignOps]{<<-}} and \verb{->>} are not allowed.}

\item{allow_right_assign}{Logical, default \code{FALSE}. If \code{TRUE}, \verb{->} and \verb{->>} are allowed.}
}
\description{
Check that \verb{<-} is always used for assignment.
}
\seealso{
\link{linters} for a complete list of linters available in lintr. \cr
\url{https://style.tidyverse.org/syntax.html#assignment-1}
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=style_linters]{style}
}
