% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backport_linter.R
\name{backport_linter}
\alias{backport_linter}
\title{Backport linter}
\usage{
backport_linter(r_version = getRversion(), except = character())
}
\arguments{
\item{r_version}{Minimum R version to test for compatibility}

\item{except}{Character vector of functions to be excluded from linting.
Use this to list explicitly defined backports, e.g. those imported from the {backports} package or manually
defined in your package.}
}
\description{
Check for usage of unavailable functions. Not reliable for testing r-devel dependencies.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=package_development_linters]{package_development}, \link[=robustness_linters]{robustness}
}
