% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backport_linter.R
\name{backport_linter}
\alias{backport_linter}
\title{Backport linter}
\usage{
backport_linter(r_version = getRversion())
}
\arguments{
\item{r_version}{Minimum R version to test for compatibility}
}
\description{
Check for usage of unavailable functions. Not reliable for testing r-devel dependencies.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=robustness_linters]{robustness}
}
