% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{best_practices_linters}
\alias{best_practices_linters}
\title{Best practices linters}
\description{
Linters checking the use of coding best practices, such as explicit typing of numeric constants.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'best_practices':
\itemize{
\item{\code{\link{absolute_path_linter}}}
\item{\code{\link{any_duplicated_linter}}}
\item{\code{\link{any_is_na_linter}}}
\item{\code{\link{boolean_arithmetic_linter}}}
\item{\code{\link{class_equals_linter}}}
\item{\code{\link{coalesce_linter}}}
\item{\code{\link{commented_code_linter}}}
\item{\code{\link{condition_call_linter}}}
\item{\code{\link{condition_message_linter}}}
\item{\code{\link{conjunct_test_linter}}}
\item{\code{\link{cyclocomp_linter}}}
\item{\code{\link{download_file_linter}}}
\item{\code{\link{empty_assignment_linter}}}
\item{\code{\link{expect_comparison_linter}}}
\item{\code{\link{expect_length_linter}}}
\item{\code{\link{expect_named_linter}}}
\item{\code{\link{expect_not_linter}}}
\item{\code{\link{expect_null_linter}}}
\item{\code{\link{expect_s3_class_linter}}}
\item{\code{\link{expect_s4_class_linter}}}
\item{\code{\link{expect_true_false_linter}}}
\item{\code{\link{expect_type_linter}}}
\item{\code{\link{fixed_regex_linter}}}
\item{\code{\link{float_comparison_linter}}}
\item{\code{\link{for_loop_index_linter}}}
\item{\code{\link{function_argument_linter}}}
\item{\code{\link{function_return_linter}}}
\item{\code{\link{if_switch_linter}}}
\item{\code{\link{ifelse_censor_linter}}}
\item{\code{\link{implicit_assignment_linter}}}
\item{\code{\link{implicit_integer_linter}}}
\item{\code{\link{is_numeric_linter}}}
\item{\code{\link{length_levels_linter}}}
\item{\code{\link{lengths_linter}}}
\item{\code{\link{library_call_linter}}}
\item{\code{\link{list_comparison_linter}}}
\item{\code{\link{literal_coercion_linter}}}
\item{\code{\link{nonportable_path_linter}}}
\item{\code{\link{nrow_subset_linter}}}
\item{\code{\link{nzchar_linter}}}
\item{\code{\link{object_overwrite_linter}}}
\item{\code{\link{outer_negation_linter}}}
\item{\code{\link{paste_linter}}}
\item{\code{\link{pipe_return_linter}}}
\item{\code{\link{print_linter}}}
\item{\code{\link{redundant_equals_linter}}}
\item{\code{\link{redundant_ifelse_linter}}}
\item{\code{\link{regex_subset_linter}}}
\item{\code{\link{rep_len_linter}}}
\item{\code{\link{routine_registration_linter}}}
\item{\code{\link{scalar_in_linter}}}
\item{\code{\link{seq_linter}}}
\item{\code{\link{sort_linter}}}
\item{\code{\link{stopifnot_all_linter}}}
\item{\code{\link{system_file_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
\item{\code{\link{terminal_close_linter}}}
\item{\code{\link{undesirable_function_linter}}}
\item{\code{\link{undesirable_operator_linter}}}
\item{\code{\link{unnecessary_lambda_linter}}}
\item{\code{\link{unnecessary_nesting_linter}}}
\item{\code{\link{unnecessary_placeholder_linter}}}
\item{\code{\link{unreachable_code_linter}}}
\item{\code{\link{unused_import_linter}}}
\item{\code{\link{vector_logic_linter}}}
\item{\code{\link{yoda_test_linter}}}
}
}
