% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brace_linter.R
\name{brace_linter}
\alias{brace_linter}
\title{Brace linter}
\usage{
brace_linter(allow_single_line = FALSE, c_style_braces = FALSE)
}
\arguments{
\item{allow_single_line}{if \code{TRUE}, allow an open and closed curly pair on the same line.}
}
\description{
Perform various style checks related to placement and spacing of curly braces:
}
\details{
\itemize{
\item Curly braces are on their own line unless they are followed by an \verb{else}.
}
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
