% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commas_linter.R
\name{commas_linter}
\alias{commas_linter}
\title{Commas linter}
\usage{
commas_linter()
}
\description{
Check that all commas are followed by spaces, but do not have spaces before them.
}
\seealso{
\link{linters} for a complete list of linters available in lintr. \cr
\url{https://style.tidyverse.org/syntax.html#commas}
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
