% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commas_linter.R
\name{commas_linter}
\alias{commas_linter}
\title{Commas linter}
\usage{
commas_linter()
}
\description{
Check that all commas are followed by spaces, but do not have spaces before them.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
