% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_message_linter.R
\name{condition_message_linter}
\alias{condition_message_linter}
\title{Block usage of \code{paste()} and \code{paste0()} with messaging functions using \code{...}}
\usage{
condition_message_linter()
}
\description{
\code{stop(paste0(...))} is strictly redundant -- \code{stop(...)} is equivalent.
\code{stop(...)} is also preferable to \code{stop(paste(...))}. The same applies to
all default condition functions, i.e., \code{\link[=stop]{stop()}}, \code{\link[=warning]{warning()}}, \code{\link[=message]{message()}},
and \code{\link[=packageStartupMessage]{packageStartupMessage()}}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}
}
