% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{configurable_linters}
\alias{configurable_linters}
\title{Configurable linters}
\description{
Generic linters which support custom configuration to your needs.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'configurable':
\itemize{
\item{\code{\link{absolute_path_linter}}}
\item{\code{\link{assignment_linter}}}
\item{\code{\link{backport_linter}}}
\item{\code{\link{brace_linter}}}
\item{\code{\link{commas_linter}}}
\item{\code{\link{condition_call_linter}}}
\item{\code{\link{conjunct_test_linter}}}
\item{\code{\link{consecutive_mutate_linter}}}
\item{\code{\link{cyclocomp_linter}}}
\item{\code{\link{decimal_fraction_linter}}}
\item{\code{\link{duplicate_argument_linter}}}
\item{\code{\link{fixed_regex_linter}}}
\item{\code{\link{if_not_else_linter}}}
\item{\code{\link{if_switch_linter}}}
\item{\code{\link{implicit_assignment_linter}}}
\item{\code{\link{implicit_integer_linter}}}
\item{\code{\link{indentation_linter}}}
\item{\code{\link{infix_spaces_linter}}}
\item{\code{\link{library_call_linter}}}
\item{\code{\link{line_length_linter}}}
\item{\code{\link{missing_argument_linter}}}
\item{\code{\link{namespace_linter}}}
\item{\code{\link{nested_pipe_linter}}}
\item{\code{\link{nonportable_path_linter}}}
\item{\code{\link{object_length_linter}}}
\item{\code{\link{object_name_linter}}}
\item{\code{\link{object_overwrite_linter}}}
\item{\code{\link{object_usage_linter}}}
\item{\code{\link{paste_linter}}}
\item{\code{\link{pipe_consistency_linter}}}
\item{\code{\link{quotes_linter}}}
\item{\code{\link{redundant_ifelse_linter}}}
\item{\code{\link{return_linter}}}
\item{\code{\link{scalar_in_linter}}}
\item{\code{\link{semicolon_linter}}}
\item{\code{\link{string_boundary_linter}}}
\item{\code{\link{todo_comment_linter}}}
\item{\code{\link{trailing_whitespace_linter}}}
\item{\code{\link{undesirable_function_linter}}}
\item{\code{\link{undesirable_operator_linter}}}
\item{\code{\link{unnecessary_concatenation_linter}}}
\item{\code{\link{unnecessary_lambda_linter}}}
\item{\code{\link{unnecessary_nesting_linter}}}
\item{\code{\link{unreachable_code_linter}}}
\item{\code{\link{unused_import_linter}}}
}
}
