% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{configurable_linters}
\alias{configurable_linters}
\title{Configurable linters}
\description{
Generic linters which support custom configuration to your needs.
}
\section{Linters}{
The following linters are tagged with 'configurable':
\itemize{
\item{\code{\link{absolute_path_linter}}}
\item{\code{\link{backport_linter}}}
\item{\code{\link{closed_curly_linter}}}
\item{\code{\link{cyclocomp_linter}}}
\item{\code{\link{duplicate_argument_linter}}}
\item{\code{\link{line_length_linter}}}
\item{\code{\link{missing_argument_linter}}}
\item{\code{\link{namespace_linter}}}
\item{\code{\link{nonportable_path_linter}}}
\item{\code{\link{object_length_linter}}}
\item{\code{\link{object_name_linter}}}
\item{\code{\link{open_curly_linter}}}
\item{\code{\link{semicolon_terminator_linter}}}
\item{\code{\link{todo_comment_linter}}}
\item{\code{\link{undesirable_function_linter}}}
\item{\code{\link{undesirable_operator_linter}}}
}
}
