% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{consistency_linters}
\alias{consistency_linters}
\title{Consistency linters}
\description{
Linters checking enforcing a consistent alternative if there are multiple syntactically valid ways to write
something.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'consistency':
\itemize{
\item{\code{\link{assignment_linter}}}
\item{\code{\link{class_equals_linter}}}
\item{\code{\link{implicit_integer_linter}}}
\item{\code{\link{no_tab_linter}}}
\item{\code{\link{object_name_linter}}}
\item{\code{\link{seq_linter}}}
\item{\code{\link{single_quotes_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
}
}
