% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{correctness_linters}
\alias{correctness_linters}
\title{Correctness linters}
\description{
Linters highlighting possible programming mistakes, such as unused variables.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'correctness':
\itemize{
\item{\code{\link{duplicate_argument_linter}}}
\item{\code{\link{equals_na_linter}}}
\item{\code{\link{missing_argument_linter}}}
\item{\code{\link{namespace_linter}}}
\item{\code{\link{non_integer_comparison_linter}}}
\item{\code{\link{object_usage_linter}}}
\item{\code{\link{package_hooks_linter}}}
\item{\code{\link{sprintf_linter}}}
}
}
