% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclocomp_linter.R
\name{cyclocomp_linter}
\alias{cyclocomp_linter}
\title{Cyclomatic complexity linter}
\usage{
cyclocomp_linter(complexity_limit = 15L)
}
\arguments{
\item{complexity_limit}{expressions with a cyclomatic complexity higher than
this are linted, defaults to 15. See \code{\link[cyclocomp]{cyclocomp}}.}
}
\description{
Check for overly complicated expressions. See \code{\link[cyclocomp]{cyclocomp}}.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
