% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_fraction_linter.R
\name{decimal_fraction_linter}
\alias{decimal_fraction_linter}
\title{Decimal fraction integer linter}
\usage{
decimal_fraction_linter()
}
\description{
Check for use of \verb{as.integer(x * <float>)} and suggest an integer arithmetic
equivalent like \code{as.integer((x * 1461L)\%/\%4L)}.
}
\examples{
# will produce a lint
lint(
  text = "as.integer(x * 365.25)",
  linters = decimal_fraction_linter()
)

lint(
  text = "as.integer(365.25 * x)",
  linters = decimal_fraction_linter()
)

lint(
  text = "as.integer(x * 0.1)",
  linters = decimal_fraction_linter()
)

lint(
  text = "as.integer(x / 0.13)",
  linters = decimal_fraction_linter()
)

# okay
lint(
  text = "as.integer(x * 365)",
  linters = decimal_fraction_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=efficiency_linters]{efficiency}, \link[=robustness_linters]{robustness}
}
