% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{default_linters}
\alias{default_linters}
\title{Default linters to use}
\format{An object of class \code{list} of length 17.}
\usage{
default_linters
}
\description{
Default linters to use
}
\keyword{datasets}
