% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{default_linters}
\alias{default_linters}
\title{Default linters to use}
\format{\preformatted{List of 17
 $ assignment_linter             :
 $ single_quotes_linter          :
 $ absolute_paths_linter         :
 $ no_tab_linter                 :
 $ line_length_linter            :
 $ commas_linter                 :
 $ infix_spaces_linter           :
 $ spaces_left_parentheses_linter:
 $ spaces_inside_linter          :
 $ open_curly_linter             :
 $ closed_curly_linter           :
 $ object_camel_case_linter      :
 $ object_multiple_dots_linter   :
 $ object_length_linter          :
 $ object_usage_linter           :
 $ trailing_whitespace_linter    :
 $ trailing_blank_lines_linter   :
}}
\usage{
default_linters
}
\description{
Default linters to use
}
\keyword{datasets}

