% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{default_linters}
\alias{default_linters}
\title{Default linters}
\format{
An object of class \code{list} of length 23.
}
\usage{
default_linters
}
\description{
List of default linters for \code{\link{lint}}. Use \code{\link{with_defaults}} to customize it.
}
\keyword{datasets}
