% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{default_settings}
\alias{default_settings}
\title{Default lintr settings}
\format{\preformatted{List of 9
 $ linters        :List of 17
  ..$ assignment_linter             :
  ..$ single_quotes_linter          :
  ..$ absolute_paths_linter         :
  ..$ no_tab_linter                 :
  ..$ line_length_linter            :
  ..$ commas_linter                 :
  ..$ infix_spaces_linter           :
  ..$ spaces_left_parentheses_linter:
  ..$ spaces_inside_linter          :
  ..$ open_curly_linter             :
  ..$ closed_curly_linter           :
  ..$ camel_case_linter             :
  ..$ multiple_dots_linter          :
  ..$ object_length_linter          :
  ..$ object_usage_linter           :
  ..$ trailing_whitespace_linter    :
  ..$ trailing_blank_lines_linter   :
 $ exclude        :Class 'regex'  chr "#[[:space:]]*nolint"
 $ exclude_start  :Class 'regex'  chr "#[[:space:]]*nolint start"
 $ exclude_end    :Class 'regex'  chr "#[[:space:]]*nolint end"
 $ exclusions     : list()
 $ cache_directory: chr "~/.R/lintr_cache"
 $ comment_token  : chr "7713914a3598997a0887b075dcc715849bf1a3ce"
 $ comment_bot    : logi TRUE
 $ error_on_lint  : logi FALSE
}}
\usage{
default_settings
}
\description{
Default lintr settings
}
\seealso{
\code{\link{read_settings}}, \code{\link{default_linters}}
}
\keyword{datasets}

