% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{default_settings}
\alias{default_settings}
\title{Default lintr settings}
\format{
An object of class \code{list} of length 12.
}
\usage{
default_settings
}
\description{
Default lintr settings
}
\seealso{
\code{\link{read_settings}}, \code{\link{default_linters}}
}
\keyword{datasets}
