% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{all_undesirable_functions}
\alias{all_undesirable_functions}
\alias{default_undesirable_functions}
\alias{all_undesirable_operators}
\alias{default_undesirable_operators}
\title{Default undesirable functions and operators}
\format{
A named list of character strings.

An object of class \code{list} of length 21.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.
}
\usage{
all_undesirable_functions

default_undesirable_functions

all_undesirable_operators

default_undesirable_operators
}
\description{
Lists of function names and operators for \code{\link[=undesirable_function_linter]{undesirable_function_linter()}} and \code{\link[=undesirable_operator_linter]{undesirable_operator_linter()}}.
There is a list for the default elements and another that contains all available elements.
Use \code{\link[=modify_defaults]{modify_defaults()}} to produce a custom list.
}
\keyword{datasets}
