% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{deprecated_linters}
\alias{deprecated_linters}
\title{Deprecated linters}
\description{
Linters that are deprecated and provided for backwards compatibility only.
These linters will be excluded from \code{\link[=linters_with_tags]{linters_with_tags()}} by default.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'deprecated':
\itemize{
\item{\code{\link{consecutive_stopifnot_linter}}}
\item{\code{\link{no_tab_linter}}}
\item{\code{\link{single_quotes_linter}}}
\item{\code{\link{unneeded_concatenation_linter}}}
}
}
