% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{deprecated_linters}
\alias{deprecated_linters}
\title{Deprecated linters}
\description{
Linters that are deprecated and provided for backwards compatiblility only.
These linters will be excluded from \code{linters_with_tags()} by default.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'deprecated':
\itemize{
\item{\code{\link{closed_curly_linter}}}
\item{\code{\link{open_curly_linter}}}
\item{\code{\link{paren_brace_linter}}}
\item{\code{\link{semicolon_terminator_linter}}}
}
}
