% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_argument_linter.R
\name{duplicate_argument_linter}
\alias{duplicate_argument_linter}
\title{Duplicate argument linter}
\usage{
duplicate_argument_linter(except = character())
}
\arguments{
\item{except}{a character vector of function names as exceptions.}
}
\description{
Check for duplicate arguments in function calls. Some of these cases will be
blocked by R's syntax itself (e.g. \code{mean(x = 1:5, x = 2:3)}), but even when
such usage \emph{is} allowed, it is rarely desirable and should be used sparingly
(e.g. \code{data.frame(x = 1, x = 2)}).
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=configurable_linters]{configurable}, \link[=correctness_linters]{correctness}
}
