% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_lint.R
\name{expect_lint}
\alias{expect_lint}
\title{Lint expectation}
\usage{
expect_lint(content, checks, ..., file = NULL)
}
\arguments{
\item{content}{a character vector for the file content to be linted.}

\item{checks}{checks to be performed:
\describe{
  \item{NULL}{check that no lints are returned.}
  \item{unnamed vector or regex object}{check if the returned lint's message matches the given
  regular expression.}
  \item{named-vector}{check if the returned lint's fields match the named fields. Accepted fields
  are the same as those taken by \code{\link{Lint}}.}
  \item{list of vectors}{check if the returned lints matches (use if more than one lint is
  returned).}
}}

\item{...}{arguments passed to \code{\link{lint}}, e.g. the linters or cache to use.}

\item{file}{if not \code{NULL}, read content from the specified file rather than from \code{content}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Lint expectation
}
\examples{
# no expected lint
expect_lint("a", NULL, trailing_blank_lines_linter)

# one expected lint
expect_lint("a\\n", "superfluous", trailing_blank_lines_linter)
expect_lint("a\\n", c(message="superfluous", line_number=2), trailing_blank_lines_linter)

# several expected lints
expect_lint("a\\n\\n", list("superfluous", "superfluous"), trailing_blank_lines_linter)
expect_lint(
  "a\\n\\n",
  list(c(message="superfluous", line_number=2), c(message="superfluous", line_number=3)),
  trailing_blank_lines_linter)
}
