% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/expect_lint.R
\name{expect_lint}
\alias{expect_lint}
\title{Lint expectation}
\usage{
expect_lint(content, checks, ...)
}
\arguments{
\item{content}{the file content to be linted}

\item{checks}{a list of named vectors of checks to be performed.  Performs
different checks depending on the value of checks.
\itemize{
  \item NULL check if the lint returns no lints.
  \item unnamed-vector check if the lint's message matches the value.
  \item named-vector check if the lint's field matches the named field.
  \item list-vectors check if the given lint matches (use if more than one lint is returned for the content)
}}

\item{...}{one or more linters to use for the check}
}
\description{
Lint expectation
}

