% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expect_lint.R
\name{expect_lint_free}
\alias{expect_lint_free}
\title{Lint free expectation}
\usage{
expect_lint_free(content, ...)
}
\arguments{
\item{content}{the file content to be linted}

\item{...}{one or more linters to use for the check}
}
\description{
Lint free expectation
}

