% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_lint.R
\name{expect_lint_free}
\alias{expect_lint_free}
\title{Test that the package is lint free}
\usage{
expect_lint_free(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{lint_package}}}
}
\description{
This function is a thin wrapper around lint_package that simply tests there are no
lints in the package.  It can be used to ensure that your tests fail if the package
contains lints.
}
