% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_s3_class_linter.R
\name{expect_s3_class_linter}
\alias{expect_s3_class_linter}
\title{Require usage of \code{expect_s3_class()}}
\usage{
expect_s3_class_linter()
}
\description{
\code{\link[testthat:inheritance-expectations]{testthat::expect_s3_class()}} exists specifically for testing the class
of S3 objects. \code{\link[testthat:equality-expectations]{testthat::expect_equal()}}, \code{\link[testthat:equality-expectations]{testthat::expect_identical()}},
and \code{\link[testthat:logical-expectations]{testthat::expect_true()}} can also be used for such tests,
but it is better to use the tailored function instead.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=package_development_linters]{package_development}
}
