% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_operator_linter.R
\name{extraction_operator_linter}
\alias{extraction_operator_linter}
\title{Extraction operator linter}
\usage{
extraction_operator_linter()
}
\description{
Check that the `[[` operator is used when extracting a single element from an object, not `[` (subsetting) nor `$`
(interactive use).
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=style_linters]{style}
}
