% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floating_point_comparison_linter.R
\name{float_comparison_linter}
\alias{float_comparison_linter}
\title{Floating point comparison linter}
\usage{
float_comparison_linter()
}
\description{
Comparing floating-point values for equality is fragile and can be affected
by rounding error. Prefer checking whether the absolute difference is within
an epsilon tolerance.
}
\examples{
# will produce lints
lint(
  text = "x == 3.0",
  linters = float_comparison_linter()
)

lint(
  text = "y == 1/10",
  linters = float_comparison_linter()
)

lint(
  text = "3.0 == x",
  linters = float_comparison_linter()
)

# okay
lint(
  text = "x == 3L",
  linters = float_comparison_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=correctness_linters]{correctness}
}
