% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_left_parentheses.R
\name{function_left_parentheses_linter}
\alias{function_left_parentheses_linter}
\title{Function left parentheses linter}
\usage{
function_left_parentheses_linter()
}
\description{
Check that all left parentheses in a function call do not have spaces before them.
}
\seealso{
\link{linters} for a complete list of linters available in lintr. \cr
\url{https://style.tidyverse.org/syntax.html#parentheses}
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
