% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_expressions.R
\name{get_source_expressions}
\alias{get_source_expressions}
\title{Parsed sourced file from a filename}
\usage{
get_source_expressions(filename)
}
\arguments{
\item{filename}{The name of the file to be parsed.}
}
\value{
A list with three elements
\describe{
\item{expressions}{A list of lists. One element for each top-level expression in the file. Each of these lists is
\describe{
\item{filename}{character. Filename of the expression}
\item{line}{integer. Line number for the start of the expression}
\item{column}{integer. Column number for the start of the expression}
\item{lines}{named character vector, where the values are the text of each line, and the names the line numbers}
\item{parsed_content}{The data frame returned by \code{\link[utils]{getParseData}}}
\item{xml_parsed_content}{An object of class \code{c("xml_document", "xml_node")}, which is the result of calling \code{\link[xml2]{read_xml}} on the XML representation of the parsed expression generated by output of \code{\link[xmlparsedata]{xml_parse_data}}.}
\item{content}{A character vector of length 1 containing the text representation of the code.}
\item{find_line}{function}
\item{find_column}{function}
}
}
\item{error}{If a parse error a \code{"lint"} object containing the error, otherwise \code{NULL}.}
\item{lines}{character vector containing the text of file split by line}
}
}
\description{
Parse a file and generate an object that is given as input to each linter.
}
