% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{gitlab_output}
\alias{gitlab_output}
\title{Gitlab Report for lint results}
\usage{
gitlab_output(lints, filename = "lintr_results.json")
}
\arguments{
\item{lints}{The linting results}

\item{filename}{The file name of the output report}
}
\description{
Generate a report of the linting results using the \href{https://docs.gitlab.com/ci/testing/code_quality/#code-quality-report-format}{Gitlab} format.
The "severity" is set to "info" for all lint results, since lintr has no
mechanism of classifying "severity".
}
