% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{gitlab_output}
\alias{gitlab_output}
\title{GitLab Report for lint results}
\usage{
gitlab_output(lints, filename = "lintr_results.json")
}
\arguments{
\item{lints}{The linting results}

\item{filename}{The file name of the output report}
}
\description{
Generate a report of the linting results using the
\href{https://docs.gitlab.com/ci/testing/code_quality/#code-quality-report-format}{GitLab} format.
}
\details{
lintr only supports three severity types ("style", "warning", and "error")
while the GitLab format supports five ("info", "minor", "major", "critical",
and "blocker"). The types "style", "warning", and "error" are mapped to
the GitLab types "info", "major", and "blocker", respectively. The GitLab
types "minor" and "critical" are ignored.
}
