% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_integer_linter.R
\name{implicit_integer_linter}
\alias{implicit_integer_linter}
\title{Implicit integer linter}
\usage{
implicit_integer_linter()
}
\description{
Check that integers are explicitly typed using the form \code{1L} instead of \code{1}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=style_linters]{style}
}
