% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation_linter.R
\name{indentation_linter}
\alias{indentation_linter}
\title{Check that indentation is consistent}
\usage{
indentation_linter(
  indent = 2L,
  hanging_indent_style = c("tidy", "always", "never")
)
}
\arguments{
\item{indent}{Number of spaces, that a code block should be indented by relative to its parent code block.
Used for multi-line code blocks (\code{{ ... }}), function calls (\code{( ... )}) and extractions (\verb{[ ... ]}, \verb{[[ ... ]]}).
Defaults to 2.}

\item{hanging_indent_style}{Indentation style for multi-line function calls with arguments in their first line.
Defaults to tidyverse style, i.e. a block indent is used if the function call terminates with \verb{)} on a separate
line and a hanging indent if not

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# complies to "tidy" and "never"
map(x, f,
  additional_arg = 42
)

# complies to "always"
map(x, f,
    additional_arg = 42
)

# complies to "tidy" and "always"
map(x, f,
    additional_arg = 42)

# complies to "never"
map(x, f,
  additional_arg = 42)
}\if{html}{\out{</div>}}}
}
\description{
Check that indentation is consistent
}
\examples{
# will produce lints
lint(
  text = "if (TRUE) {\n1 + 1\n}",
  linters = indentation_linter()
)

lint(
  text = "if (TRUE) {\n    1 + 1\n}",
  linters = indentation_linter()
)

lint(
  text = "map(x, f,\n  additional_arg = 42\n)",
  linters = indentation_linter(hanging_indent_style = "always")
)

lint(
  text = "map(x, f,\n    additional_arg = 42)",
  linters = indentation_linter(hanging_indent_style = "never")
)

# okay
lint(
  text = "map(x, f,\n  additional_arg = 42\n)",
  linters = indentation_linter()
)

lint(
  text = "if (TRUE) {\n    1 + 1\n}",
  linters = indentation_linter(indent = 4)
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr. \cr
\url{https://style.tidyverse.org/syntax.html#indenting}
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
