% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation_linter.R
\name{indentation_linter}
\alias{indentation_linter}
\title{Check that indentation is consistent}
\usage{
indentation_linter(indent = 2L)
}
\arguments{
\item{indent}{Block indentation level, used for multi-line code blocks (\code{{ ... }}), function calls (\code{( ... )}) and
extractions (\verb{[ ... ]}, \verb{[[ ... ]]}).}
}
\description{
Check that indentation is consistent
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
