% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation_linter.R
\name{indentation_linter}
\alias{indentation_linter}
\title{Check that indentation is consistent}
\usage{
indentation_linter(indent = 2L, use_hybrid_indent = TRUE)
}
\arguments{
\item{indent}{Block indentation level, used for multi-line code blocks (\code{{ ... }}), function calls (\code{( ... )}) and
extractions (\verb{[ ... ]}, \verb{[[ ... ]]}).}

\item{use_hybrid_indent}{Require a block indent for multi-line function calls if there are only unnamed arguments
in the first line?

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# complies only with use_hybrid_indent = TRUE:
map(x, f,
  additional_arg = 42
)

# complies only with use_hybrid_indent = FALSE:
map(x, f,
    additional_arg = 42
)

# always complies
map(x, f,
    additional_arg = 42)
}\if{html}{\out{</div>}}}
}
\description{
Check that indentation is consistent
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
