% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_spaces_linter.R
\name{infix_spaces_linter}
\alias{infix_spaces_linter}
\title{Infix spaces linter}
\usage{
infix_spaces_linter()
}
\description{
Check that infix operators are surrounded by spaces.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
