% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_comparison_linter.R
\name{inner_comparison_linter}
\alias{inner_comparison_linter}
\title{Require == to be used outside of sapply() when comparing to a constant}
\usage{
inner_comparison_linter()
}
\description{
\code{sapply(x, function(xi) foo(xi) == 2)} is the same as \code{sapply(x, foo) == 2},
but harder to read and requires two passes over the vector.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=efficiency_linters]{efficiency}
}
