% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_length_linter.R
\name{line_length_linter}
\alias{line_length_linter}
\title{Line length linter}
\usage{
line_length_linter(length = 80L)
}
\arguments{
\item{length}{maximum line length allowed.}
}
\description{
Check that the line length of both comments and code is less than \code{length}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=configurable_linters]{configurable}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
