% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{lint}
\alias{lint}
\title{Lint a Project}
\usage{
lint(filename, line_number = 1L, column_number = NULL, type = "style",
  message = "", line = "", ranges = NULL)
}
\arguments{
\item{project}{Path to a project directory.}
}
\description{
Takes the set of active linters (see \code{\link{add_linter}}), and applies
them to all files within a project.
}

