% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lint.R
\name{Lint}
\alias{Lint}
\title{Create a \code{Lint} object}
\usage{
Lint(filename, line_number = 1L, column_number = NULL, type = c("style",
  "warning", "error"), message = "", line = "", ranges = NULL,
  linter = NULL)
}
\arguments{
\item{filename}{path to the source file that was linted.}

\item{line_number}{line number where the lint occurred.}

\item{column_number}{column the lint occurred.}

\item{type}{type of lint.}

\item{message}{message used to describe the lint error}

\item{line}{code source where the lint occured}

\item{ranges}{ranges on the line that should be emphasized.}

\item{linter}{name of linter that created the Lint object.}
}
\description{
Create a \code{Lint} object
}

