% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_file}
\alias{lint_file}
\alias{lint}
\title{Lint a file}
\usage{
lint(
  filename,
  linters = NULL,
  cache = FALSE,
  ...,
  parse_settings = TRUE,
  text = NULL
)
}
\arguments{
\item{filename}{the given filename to lint.}

\item{linters}{a named list of linter functions to apply see \code{\link{linters}}
for a full list of default and available linters.}

\item{cache}{given a logical, toggle caching of lint results. If passed a
character string, store the cache in this directory.}

\item{...}{additional arguments passed to \code{\link{exclude}}.}

\item{parse_settings}{whether to try and parse the settings}

\item{text}{the string content to be used as if linting \code{filename}.
If supplied, \code{filename} must be a valid path.}
}
\value{
A list of lint objects.
}
\description{
Apply one or more linters to a file and return the lints found.
}
