% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_file}
\alias{lint}
\alias{lint_file}
\title{Lint a given file}
\usage{
lint(filename, linters = NULL, cache = FALSE, ..., parse_settings = TRUE)
}
\arguments{
\item{filename}{the given filename to lint.}

\item{linters}{a list of linter functions to apply see \code{\link{linters}}
for a full list of default and available linters.}

\item{cache}{toggle caching of lint results}

\item{...}{additional arguments passed to \code{\link{exclude}}.}

\item{parse_settings}{whether to try and parse the settings}
}
\description{
Apply one or more linters to a file and return a list of lints found.
}

