% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Linter}
\alias{Linter}
\title{Create a \code{linter} closure}
\usage{
Linter(
  fun,
  name = linter_auto_name(),
  linter_level = c(NA_character_, "file", "expression")
)
}
\arguments{
\item{fun}{A function that takes a source file and returns \code{lint} objects.}

\item{name}{Default name of the Linter.
Lints produced by the linter will be labelled with \code{name} by default.}

\item{linter_level}{Which level of expression is the linter working with?
\code{"expression"} means an individual expression in \code{xml_parsed_content}, while \code{"file"} means all expressions
in the current file are available in \code{full_xml_parsed_content}.
\code{NA} means the linter will be run with both, expression-level and file-level source expressions.}
}
\value{
The same function with its class set to 'linter'.
}
\description{
Create a \code{linter} closure
}
