% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Linter}
\alias{Linter}
\title{Create a \code{linter} closure}
\usage{
Linter(fun)
}
\arguments{
\item{fun}{a function that takes a source file and returns \code{lint} objects.}
}
\value{
the same function with its class set to 'linter'.
}
\description{
Create a \code{linter} closure
}
