% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{linter}
\alias{linter}
\title{Create a Linter}
\usage{
linter(apply, takes, message, suggestion)
}
\arguments{
\item{apply}{Function that, given the content of a file, returns the indices
at which problems were found.}

\item{takes}{Function that, given a set of paths, returns the subset of
paths that this linter uses.}

\item{message}{Function that, given content and lines, returns an
informative message for the user. Typically generated with
\code{\link{make_linter_message}}.}

\item{suggestion}{String giving a prescribed fix for the linted problem.}
}
\description{
Generate a linter, which can identify errors or problematic regions in a
project.
}

