% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{absolute_paths_linter}
\alias{absolute_paths_linter}
\alias{assignment_linter}
\alias{closed_curly_linter}
\alias{commas_linter}
\alias{infix_spaces_linter}
\alias{line_length_linter}
\alias{linters}
\alias{no_tab_linter}
\alias{object_name_linter}
\alias{open_curly_linter}
\alias{single_quotes_linter}
\alias{spaces_inside_linter}
\alias{spaces_left_parentheses_linter}
\title{linters}
\usage{
absolute_paths_linter(source_file)

assignment_linter(source_file)

closed_curly_linter(source_file)

commas_linter(source_file)

infix_spaces_linter(source_file)

line_length_linter(width)

no_tab_linter(source_file)

object_name_linter(width = 20L)

open_curly_linter(source_file)

single_quotes_linter(source_file)

spaces_inside_linter(source_file)

spaces_left_parentheses_linter(source_file)
}
\arguments{
\item{source_file}{returned by \code{\link{get_source_file}}}

\item{width}{the width cutoff to use for the given linter.}
}
\description{
Available linters
}
\section{Functions}{
\itemize{
\item \code{absolute_paths_linter}: checks that no absolute paths are used.

\item \code{assignment_linter}: checks that '<-' is always used for assignment

\item \code{closed_curly_linter}: check that closed curly braces should always be on their
own line unless they follow an else.

\item \code{commas_linter}: check that all commas are followed by spaces, but do not
have spaces before them.

\item \code{infix_spaces_linter}: check that all infix operators have spaces around them.

\item \code{line_length_linter}: check the line length of both comments and code is less than
width.

\item \code{no_tab_linter}: check that only spaces are used, never tabs.

\item \code{object_name_linter}: check that objects
\enumerate{
  \item Are never camelCase
  \item Are separated by \code{_} rather than \code{.}
  \item Are not more than \code{width} characters
}

\item \code{open_curly_linter}: check that opening curly braces are never on their own
line and are always followed by a newline.

\item \code{single_quotes_linter}: checks that only single quotes are used to delimit
string contestants.

\item \code{spaces_inside_linter}: check that parentheses and square brackets do not have
spaces directly inside them.

\item \code{spaces_left_parentheses_linter}: check that all left parentheses have a space before them
unless they are in a function call.
}}

