% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{linters}
\alias{linters}
\title{Available linters}
\description{
A variety of linters is available in \pkg{lintr}. The most popular ones are readily
accessible through \code{\link[=default_linters]{default_linters()}}, though there are additional ones you may want
to use.

Within a \code{\link[=lint]{lint()}} function call, the linters in use are initialized with the provided
arguments and fed with the source file (provided by \code{\link[=get_source_expressions]{get_source_expressions()}}).

A data frame of all available linters can be retrieved using \code{\link[=available_linters]{available_linters()}}.
Documentation for linters is structured into tags to allow for easier discovery.
}
\section{Tags}{
The following tags exist:
\itemize{
\item{\link[=best_practices_linters]{best_practices} (11 linters)}
\item{\link[=common_mistakes_linters]{common_mistakes} (5 linters)}
\item{\link[=configurable_linters]{configurable} (16 linters)}
\item{\link[=consistency_linters]{consistency} (7 linters)}
\item{\link[=correctness_linters]{correctness} (7 linters)}
\item{\link[=default_linters]{default} (25 linters)}
\item{\link[=efficiency_linters]{efficiency} (4 linters)}
\item{\link[=package_development_linters]{package_development} (1 linters)}
\item{\link[=readability_linters]{readability} (21 linters)}
\item{\link[=robustness_linters]{robustness} (10 linters)}
\item{\link[=style_linters]{style} (32 linters)}
}
}
\section{Linters}{
The following linters exist:
\itemize{
\item{\code{\link{absolute_path_linter}} (tags: best_practices, configurable, robustness)}
\item{\code{\link{assignment_linter}} (tags: consistency, default, style)}
\item{\code{\link{assignment_spaces_linter}} (tags: readability, style)}
\item{\code{\link{backport_linter}} (tags: configurable, robustness)}
\item{\code{\link{closed_curly_linter}} (tags: configurable, default, readability, style)}
\item{\code{\link{commas_linter}} (tags: default, readability, style)}
\item{\code{\link{commented_code_linter}} (tags: best_practices, default, readability, style)}
\item{\code{\link{cyclocomp_linter}} (tags: best_practices, configurable, default, readability, style)}
\item{\code{\link{duplicate_argument_linter}} (tags: common_mistakes, configurable, correctness)}
\item{\code{\link{equals_na_linter}} (tags: common_mistakes, correctness, default, robustness)}
\item{\code{\link{expect_null_linter}} (tags: best_practices, package_development)}
\item{\code{\link{extraction_operator_linter}} (tags: best_practices, style)}
\item{\code{\link{function_left_parentheses_linter}} (tags: default, readability, style)}
\item{\code{\link{implicit_integer_linter}} (tags: best_practices, consistency, style)}
\item{\code{\link{infix_spaces_linter}} (tags: default, readability, style)}
\item{\code{\link{line_length_linter}} (tags: configurable, default, readability, style)}
\item{\code{\link{missing_argument_linter}} (tags: common_mistakes, configurable, correctness)}
\item{\code{\link{missing_package_linter}} (tags: common_mistakes, robustness)}
\item{\code{\link{namespace_linter}} (tags: configurable, correctness, robustness)}
\item{\code{\link{nonportable_path_linter}} (tags: best_practices, configurable, robustness)}
\item{\code{\link{no_tab_linter}} (tags: consistency, default, style)}
\item{\code{\link{object_length_linter}} (tags: configurable, default, readability, style)}
\item{\code{\link{object_name_linter}} (tags: configurable, consistency, default, style)}
\item{\code{\link{object_usage_linter}} (tags: correctness, default, readability, style)}
\item{\code{\link{open_curly_linter}} (tags: configurable, default, readability, style)}
\item{\code{\link{package_hooks_linter}} (tags: correctness, style)}
\item{\code{\link{paren_body_linter}} (tags: default, readability, style)}
\item{\code{\link{paren_brace_linter}} (tags: default, readability, style)}
\item{\code{\link{pipe_call_linter}} (tags: readability, style)}
\item{\code{\link{pipe_continuation_linter}} (tags: default, readability, style)}
\item{\code{\link{semicolon_terminator_linter}} (tags: configurable, default, readability, style)}
\item{\code{\link{seq_linter}} (tags: best_practices, consistency, default, efficiency, robustness)}
\item{\code{\link{single_quotes_linter}} (tags: consistency, default, readability, style)}
\item{\code{\link{spaces_inside_linter}} (tags: default, readability, style)}
\item{\code{\link{spaces_left_parentheses_linter}} (tags: default, readability, style)}
\item{\code{\link{sprintf_linter}} (tags: common_mistakes, correctness)}
\item{\code{\link{T_and_F_symbol_linter}} (tags: best_practices, consistency, default, readability, robustness, style)}
\item{\code{\link{todo_comment_linter}} (tags: configurable, style)}
\item{\code{\link{trailing_blank_lines_linter}} (tags: default, style)}
\item{\code{\link{trailing_whitespace_linter}} (tags: default, style)}
\item{\code{\link{undesirable_function_linter}} (tags: best_practices, configurable, efficiency, robustness, style)}
\item{\code{\link{undesirable_operator_linter}} (tags: best_practices, configurable, efficiency, robustness, style)}
\item{\code{\link{unneeded_concatenation_linter}} (tags: efficiency, readability, style)}
}
}
