% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr-deprecated.R
\name{lintr-deprecated}
\alias{lintr-deprecated}
\alias{closed_curly_linter}
\alias{open_curly_linter}
\alias{paren_brace_linter}
\alias{semicolon_terminator_linter}
\alias{unneeded_concatenation_linter}
\alias{single_quotes_linter}
\alias{consecutive_stopifnot_linter}
\alias{no_tab_linter}
\alias{extraction_operator_linter}
\alias{unnecessary_nested_if_linter}
\title{Deprecated functions in lintr}
\usage{
closed_curly_linter(allow_single_line = FALSE)

open_curly_linter(allow_single_line = FALSE)

paren_brace_linter()

semicolon_terminator_linter(semicolon = c("compound", "trailing"))

unneeded_concatenation_linter(allow_single_expression = TRUE)

single_quotes_linter()

consecutive_stopifnot_linter()

no_tab_linter()

extraction_operator_linter()

unnecessary_nested_if_linter()
}
\arguments{
\item{allow_single_line, semicolon}{Irrelevant parameters to defunct linters.}
}
\description{
These functions have been deprecated from lintr.
\itemize{
\item \code{open_curly_linter()} (use \code{\link[=brace_linter]{brace_linter()}})
\item \code{closed_curly_linter()} (use \code{brace_linter()})
\item \code{paren_brace_linter()} (use \code{brace_linter()})
\item \code{semicolon_terminator_linter()} (use \code{\link[=semicolon_linter]{semicolon_linter()}})
}
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\keyword{internal}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=deprecated_linters]{deprecated}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
