% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr-deprecated.R
\name{lintr-deprecated}
\alias{lintr-deprecated}
\alias{unneeded_concatenation_linter}
\alias{single_quotes_linter}
\alias{consecutive_stopifnot_linter}
\alias{no_tab_linter}
\alias{extraction_operator_linter}
\alias{unnecessary_nested_if_linter}
\title{Deprecated functions in lintr}
\usage{
unneeded_concatenation_linter(allow_single_expression = TRUE)

single_quotes_linter()

consecutive_stopifnot_linter()

no_tab_linter()

extraction_operator_linter()

unnecessary_nested_if_linter()
}
\description{
Deprecated functions in lintr

Unnecessary concatenation linter

Single quotes linter

Consecutive stopifnot linter

No tabs linter

Extraction operator linter

Unnecessary nested if linter
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\keyword{internal}
\section{Tags}{
\link[=deprecated_linters]{deprecated}
}
