% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\docType{package}
\name{lintr-package}
\alias{lintr}
\alias{lintr-package}
\title{Lintr}
\description{
Checks adherence to a given style, syntax errors and possible semantic issues.
Supports on the fly checking of R code edited with Emacs, Vim and Sublime Text.
}
\seealso{
\code{\link[=lint]{lint()}}, \code{\link[=lint_package]{lint_package()}}, \code{\link[=lint_dir]{lint_dir()}}, \link{linters}
}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com}

Authors:
\itemize{
  \item Florent Angly (fangly)
  \item Russ Hyde
  \item Michael Chirico
  \item Kun Ren
  \item Alexander Rosenstock (AshesITR)
}

}
\keyword{internal}
