% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lint.R
\name{lintr}
\alias{lintr}
\title{Lintr}
\description{
Checks adherence to a given style, syntax errors and possible semantic
issues.  Supports on the fly checking of R code edited with Emacs, Vim and
Sublime Text.
}
\seealso{
\code{\link{lint}}, \code{\link{lint_package}}, \code{\link{linters}}
}

