% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_argument_linter.R
\name{missing_argument_linter}
\alias{missing_argument_linter}
\title{Missing argument linter}
\usage{
missing_argument_linter(except = c("switch", "alist"))
}
\arguments{
\item{except}{a character vector of function names as exceptions.}
}
\description{
Check for missing arguments in function calls.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=configurable_linters]{configurable}, \link[=correctness_linters]{correctness}
}
