% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_package_linter.R
\name{missing_package_linter}
\alias{missing_package_linter}
\title{Missing package linter}
\usage{
missing_package_linter()
}
\description{
Check for missing packages in \code{library()}, \code{require()}, \code{loadNamespace()} and \code{requireNamespace()} calls.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=robustness_linters]{robustness}
}
