% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_linters.R
\name{nonportable_path_linter}
\alias{nonportable_path_linter}
\title{Non-portable path linter}
\usage{
nonportable_path_linter(lax = TRUE)
}
\arguments{
\item{lax}{Less stringent linting, leading to fewer false positives.}
}
\description{
Check that file.path() is used to construct safe and portable paths.
}
