% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{normalize_exclusions}
\alias{normalize_exclusions}
\title{Normalize lint exclusions}
\usage{
normalize_exclusions(x, normalize_path = TRUE, root = getwd(), pattern = NULL)
}
\arguments{
\item{x}{Exclusion specification
- A character vector of filenames or directories relative to \code{root}
- A named list of integers specifying lines to be excluded per file}

\item{normalize_path}{Should the names of the returned exclusion list be normalized paths?
If no, they will be relative to \code{root}.}

\item{root}{Base directory for relative filename resolution.}

\item{pattern}{If non-NULL, only exclude files in excluded directories if they match
\code{pattern}. Passed to \link[base]{list.files} if a directory is excluded.}
}
\description{
Normalize lint exclusions
}
\keyword{internal}
