% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_name_linters.R
\name{object_length_linter}
\alias{object_length_linter}
\title{Object length linter}
\usage{
object_length_linter(length = 30L)
}
\arguments{
\item{length}{maximum variable name length allowed.}
}
\description{
Check that object names are not too long.
}
