% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_usage_linter.R
\name{object_usage_linter}
\alias{object_usage_linter}
\title{Object usage linter}
\usage{
object_usage_linter()
}
\description{
Check that closures have the proper usage using
\code{\link[codetools]{checkUsage}}. Note that this runs
\code{\link[base]{eval}} on the code, so do not use with untrusted code.
}
\section{Tags}{
\link[=correctness_linters]{correctness}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
