% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_usage_linter.R
\name{object_usage_linter}
\alias{object_usage_linter}
\title{Object usage linter}
\usage{
object_usage_linter(interpret_glue = TRUE)
}
\arguments{
\item{interpret_glue}{If TRUE, interpret \code{\link[glue:glue]{glue::glue()}} calls to avoid false positives caused by local variables
which are only used in a glue expression.}
}
\description{
Check that closures have the proper usage using \code{\link[codetools:checkUsage]{codetools::checkUsage()}}.
Note that this runs \code{\link[base:eval]{base::eval()}} on the code, so do not use with untrusted code.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=correctness_linters]{correctness}, \link[=default_linters]{default}, \link[=executing_linters]{executing}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
