% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_hooks_linter.R
\name{package_hooks_linter}
\alias{package_hooks_linter}
\title{Package hooks linter}
\usage{
package_hooks_linter()
}
\description{
Check various common "gotchas" in [.onLoad()], [.onAttach()], [.Last.lib()], and [.onDetach()]
  namespace hooks that will cause `R CMD check` issues.
}
