% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_hooks_linter.R
\name{package_hooks_linter}
\alias{package_hooks_linter}
\title{Package hooks linter}
\usage{
package_hooks_linter()
}
\description{
Check various common "gotchas" in \code{\link[=.onLoad]{.onLoad()}}, \code{\link[=.onAttach]{.onAttach()}}, \code{\link[=.Last.lib]{.Last.lib()}}, and \code{\link[=.onDetach]{.onDetach()}}
namespace hooks that will cause \verb{R CMD check} issues.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=correctness_linters]{correctness}, \link[=style_linters]{style}
}
