% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_consistency_linter.R
\name{pipe_consistency_linter}
\alias{pipe_consistency_linter}
\title{Pipe consistency linter}
\usage{
pipe_consistency_linter()
}
\description{
Check that all pipes are consistent (either all \%>\% or all |>).
}
\details{
Checks consistency across an entire file, not just a single
expression.
}
\examples{
# will produce lints
lint(
  text = "1:3 |> mean() \%>\% as.character()",
  linters = pipe_consistency_linter()
)

# okay
lint(
  text = "1:3 \%>\% mean() \%>\% as.character()",
  linters = pipe_consistency_linter()
)

lint(
 text = "1:3 |> mean() |> as.character()",
 linters = pipe_consistency_linter()
)
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=readability_linters]{readability}, \link[=style_linters]{style}
}
