% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_continuation_linter.R
\name{pipe_continuation_linter}
\alias{pipe_continuation_linter}
\title{Pipe continuation linter}
\usage{
pipe_continuation_linter()
}
\description{
Check that each step in a pipeline is on a new line, or the entire pipe fits on one line.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
}
