% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{pkg_testthat_linters}
\alias{pkg_testthat_linters}
\title{Testthat linters}
\description{
Linters encouraging best practices within testthat suites.
}
\seealso{
\itemize{
\item \link{linters} for a complete list of linters available in lintr.
\item \url{https://testthat.r-lib.org}
\item \url{https://r-pkgs.org/testing-basics.html}
}
}
\section{Linters}{
The following linters are tagged with 'pkg_testthat':
\itemize{
\item{\code{\link{conjunct_test_linter}}}
\item{\code{\link{expect_comparison_linter}}}
\item{\code{\link{expect_identical_linter}}}
\item{\code{\link{expect_length_linter}}}
\item{\code{\link{expect_named_linter}}}
\item{\code{\link{expect_not_linter}}}
\item{\code{\link{expect_null_linter}}}
\item{\code{\link{expect_s3_class_linter}}}
\item{\code{\link{expect_s4_class_linter}}}
\item{\code{\link{expect_true_false_linter}}}
\item{\code{\link{expect_type_linter}}}
\item{\code{\link{yoda_test_linter}}}
}
}
