% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{read_settings}
\alias{read_settings}
\title{Read lintr settings}
\usage{
read_settings(filename)
}
\arguments{
\item{filename}{source file to be linted}
}
\description{
Lintr searches for settings for a given file in the following order.
\enumerate{
  \item options defined as \code{linter.setting}.
  \item \code{linter_file} in the same directory
  \item \code{linter_file} in the project directory
  \item \code{\link{default_settings}}
}
}
\details{
The default linter_file name is \code{.lintr} but it can be changed with option
\code{lintr.linter_file}.  This file is a dcf file, see \code{\link{read.dcf}} for details.
}

