% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{readability_linters}
\alias{readability_linters}
\title{Readability linters}
\description{
Linters highlighting readability issues, such as missing whitespace.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'readability':
\itemize{
\item{\code{\link{closed_curly_linter}}}
\item{\code{\link{commas_linter}}}
\item{\code{\link{commented_code_linter}}}
\item{\code{\link{cyclocomp_linter}}}
\item{\code{\link{expect_length_linter}}}
\item{\code{\link{expect_named_linter}}}
\item{\code{\link{expect_not_linter}}}
\item{\code{\link{expect_true_false_linter}}}
\item{\code{\link{function_left_parentheses_linter}}}
\item{\code{\link{infix_spaces_linter}}}
\item{\code{\link{line_length_linter}}}
\item{\code{\link{object_length_linter}}}
\item{\code{\link{object_usage_linter}}}
\item{\code{\link{open_curly_linter}}}
\item{\code{\link{paren_body_linter}}}
\item{\code{\link{paren_brace_linter}}}
\item{\code{\link{pipe_call_linter}}}
\item{\code{\link{pipe_continuation_linter}}}
\item{\code{\link{semicolon_terminator_linter}}}
\item{\code{\link{single_quotes_linter}}}
\item{\code{\link{spaces_inside_linter}}}
\item{\code{\link{spaces_left_parentheses_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
\item{\code{\link{unneeded_concatenation_linter}}}
\item{\code{\link{yoda_test_linter}}}
}
}
