% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_linter.R
\name{repeat_linter}
\alias{repeat_linter}
\title{Repeat linter}
\usage{
repeat_linter()
}
\description{
Check that \verb{while (TRUE)} is not used for infinite loops.
}
\examples{
# will produce lints
lint(
  text = "while (TRUE) { }",
  linters = repeat_linter()
)


# okay
lint(
  text = "repeat { }",
  linters = repeat_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=readability_linters]{readability}, \link[=style_linters]{style}
}
