% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_linter.R
\name{return_linter}
\alias{return_linter}
\title{Return linter}
\usage{
return_linter(use_implicit_returns = TRUE)
}
\arguments{
\item{use_implicit_returns}{Whether to use implicit or explicit returns}
}
\description{
This linter checks for explicit \code{\link[=return]{return()}} at the end of a function
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=style_linters]{style}
}
