% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_tag_docs.R
\name{robustness_linters}
\alias{robustness_linters}
\title{Robustness linters}
\description{
Linters highlighting code robustness issues, such as possibly wrong edge case behavior.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Linters}{
The following linters are tagged with 'robustness':
\itemize{
\item{\code{\link{absolute_path_linter}}}
\item{\code{\link{all_equal_linter}}}
\item{\code{\link{backport_linter}}}
\item{\code{\link{class_equals_linter}}}
\item{\code{\link{download_file_linter}}}
\item{\code{\link{equals_na_linter}}}
\item{\code{\link{for_loop_index_linter}}}
\item{\code{\link{missing_package_linter}}}
\item{\code{\link{namespace_linter}}}
\item{\code{\link{nonportable_path_linter}}}
\item{\code{\link{object_overwrite_linter}}}
\item{\code{\link{routine_registration_linter}}}
\item{\code{\link{sample_int_linter}}}
\item{\code{\link{seq_linter}}}
\item{\code{\link{strings_as_factors_linter}}}
\item{\code{\link{T_and_F_symbol_linter}}}
\item{\code{\link{terminal_close_linter}}}
\item{\code{\link{undesirable_function_linter}}}
\item{\code{\link{undesirable_operator_linter}}}
}
}
