% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_linter.R
\name{seq_linter}
\alias{seq_linter}
\title{Sequence linter}
\usage{
seq_linter()
}
\description{
Check for \code{1:length(...)}, \code{1:nrow(...)}, \code{1:ncol(...)}, \code{1:NROW(...)} and \code{1:NCOL(...)} expressions in base-R.
Additionally check \code{1:n()} (from dplyr) and \code{1:.N} (from data.table).
These often cause bugs when the right-hand side is zero.
It is safer to use \code{\link[base:seq]{base::seq_len()}} or \code{\link[base:seq]{base::seq_along()}} instead.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=efficiency_linters]{efficiency}, \link[=robustness_linters]{robustness}
}
