% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces_inside_linter.R
\name{spaces_inside_linter}
\alias{spaces_inside_linter}
\title{Spaces inside linter}
\usage{
spaces_inside_linter()
}
\description{
Check that parentheses and square brackets do not have spaces directly inside them.
}
