% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_linter.R
\name{sprintf_linter}
\alias{sprintf_linter}
\title{Require correct \code{sprintf()} calls}
\usage{
sprintf_linter()
}
\description{
Check for an inconsistent number of arguments or arguments with incompatible types
(for literal arguments) in \code{sprintf()} calls.
}
\examples{
# will produce lints
lint(
  text = "sprintf('hello \%s \%s \%d', x, y)",
  linters = sprintf_linter()
)

# okay
lint(
  text = "sprintf('hello \%s \%s \%d', x, y, z)",
  linters = sprintf_linter()
)

lint(
  text = "sprintf('hello \%s \%s \%d', x, y, ...)",
  linters = sprintf_linter()
)

}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=correctness_linters]{correctness}
}
