% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_linter.R
\name{sprintf_linter}
\alias{sprintf_linter}
\title{\code{sprintf} linter}
\usage{
sprintf_linter()
}
\description{
Check for inconsistent number of arguments or arguments with incompatible types in \code{sprintf} calls.
}
