% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_linter.R
\name{sprintf_linter}
\alias{sprintf_linter}
\title{Require correct usage of \code{sprintf()}}
\usage{
sprintf_linter()
}
\description{
Check for an inconsistent number of arguments or arguments with incompatible types (for literal arguments) in
\code{sprintf()} calls.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=correctness_linters]{correctness}
}
