% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_linter.R
\name{sprintf_linter}
\alias{sprintf_linter}
\title{\code{sprintf} linter}
\usage{
sprintf_linter()
}
\description{
Check for an inconsistent number of arguments or arguments with incompatible types (for literal arguments) in
\code{sprintf} calls.
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=correctness_linters]{correctness}
}
\section{Tags}{
\link[=common_mistakes_linters]{common_mistakes}, \link[=correctness_linters]{correctness}
}
