% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_F_symbol_linter.R
\name{T_and_F_symbol_linter}
\alias{T_and_F_symbol_linter}
\title{\code{T} and \code{F} symbol linter}
\usage{
T_and_F_symbol_linter()
}
\description{
Avoid the symbols \code{T} and \code{F} (for \code{TRUE} and \code{FALSE}).
}
