% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_F_symbol_linter.R
\name{T_and_F_symbol_linter}
\alias{T_and_F_symbol_linter}
\title{\code{T} and \code{F} symbol linter}
\usage{
T_and_F_symbol_linter()
}
\description{
Avoid the symbols \code{T} and \code{F} (for \code{TRUE} and \code{FALSE}).
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=consistency_linters]{consistency}, \link[=default_linters]{default}, \link[=readability_linters]{readability}, \link[=robustness_linters]{robustness}, \link[=style_linters]{style}
}
