% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_linters.R
\name{todo_comment_linter}
\alias{todo_comment_linter}
\title{TODO comment linter}
\usage{
todo_comment_linter(todo = c("todo", "fixme"))
}
\arguments{
\item{todo}{Vector of strings that identify TODO comments.}
}
\description{
Check that the source contains no TODO comments (case-insensitive).
}
