% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailing_blank_lines_linter.R
\name{trailing_blank_lines_linter}
\alias{trailing_blank_lines_linter}
\title{Trailing blank lines linter}
\usage{
trailing_blank_lines_linter()
}
\description{
Check that there are no trailing blank lines in source files.
}
