% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailing_whitespace_linter.R
\name{trailing_whitespace_linter}
\alias{trailing_whitespace_linter}
\title{Trailing whitespace linter}
\usage{
trailing_whitespace_linter(allow_empty_lines = FALSE, allow_in_strings = TRUE)
}
\arguments{
\item{allow_empty_lines}{Suppress lints for lines that contain only whitespace.}

\item{allow_in_strings}{Suppress lints for trailing whitespace in string constants.}
}
\description{
Check that there are no space characters at the end of source lines.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=default_linters]{default}, \link[=style_linters]{style}
}
