% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undesirable_function_linter.R
\name{undesirable_function_linter}
\alias{undesirable_function_linter}
\title{Undesirable function linter}
\usage{
undesirable_function_linter(
  fun = default_undesirable_functions,
  symbol_is_undesirable = TRUE
)
}
\arguments{
\item{fun}{Named character vector. \code{names(fun)} correspond to undesirable functions,
while the values give a description of why the function is undesirable.
If \code{NA}, no additional information is given in the lint message. Defaults to
\link{default_undesirable_functions}. To make small customizations to this list,
use \code{\link[=modify_defaults]{modify_defaults()}}.}

\item{symbol_is_undesirable}{Whether to consider the use of an undesirable function name as a symbol undesirable
or not.}
}
\description{
Report the use of undesirable functions (e.g. \code{\link[base:function]{base::return()}}, \code{\link[base:options]{base::options()}}, or
\code{\link[base:lapply]{base::sapply()}}) and suggest an alternative.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=configurable_linters]{configurable}, \link[=efficiency_linters]{efficiency}, \link[=robustness_linters]{robustness}, \link[=style_linters]{style}
}
