% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undesirable_function_linter.R
\name{undesirable_function_linter}
\alias{undesirable_function_linter}
\title{Undesirable function linter}
\usage{
undesirable_function_linter(
  fun = default_undesirable_functions,
  symbol_is_undesirable = TRUE
)
}
\arguments{
\item{fun}{Named character vector, where the names are the names of the
undesirable functions, and the values are the text for the alternative
function to use (or \code{NA}).}

\item{symbol_is_undesirable}{Whether to consider the use of an undesirable
function name as a symbol undesirable or not.}
}
\description{
Report the use of undesirable functions, e.g. \code{return}, \code{options}, or \code{sapply} and suggest an
alternative.
}
