% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undesirable_operator_linter.R
\name{undesirable_operator_linter}
\alias{undesirable_operator_linter}
\title{Undesirable operator linter}
\usage{
undesirable_operator_linter(op = default_undesirable_operators)
}
\arguments{
\item{op}{Named character vector, where the names are the names of the undesirable operators,
and the values are the text for the alternative operator to use (or \code{NA}).}
}
\description{
Report the use of undesirable operators, e.g. \code{`:::`} or \code{`<<-`} and suggest an alternative.
}
