% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnecessary_nesting_linter.R
\name{unnecessary_nesting_linter}
\alias{unnecessary_nesting_linter}
\title{Block instances of excessive nesting}
\usage{
unnecessary_nesting_linter()
}
\description{
See \href{https://goto.google.com/c-readability-advice#avoid-unnecessary-nesting-by-using-early-returns.}{go/c-readability-advice#avoid-unnecessary-nesting-by-using-early-returns.}
This linter covers instances of excessive nesting in R code.
}
\details{
NB: This does not apply to in-line if/else statements, which, per
\href{https://goto.google.com/rstyle#inline-statements}{go/rstyle#inline-statements}, shouldn't be in-line anyway.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=consistency_linters]{consistency}, \link[=readability_linters]{readability}
}
