% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unneeded_concatenation_linter.R
\name{unneeded_concatenation_linter}
\alias{unneeded_concatenation_linter}
\title{Unneeded concatenation linter}
\usage{
unneeded_concatenation_linter()
}
\description{
Check that the \code{c} function is not used without arguments nor with a single constant.
}
