% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unreachable_code_linter.R
\name{unreachable_code_linter}
\alias{unreachable_code_linter}
\title{Block unreachable code and comments following return statements}
\usage{
unreachable_code_linter()
}
\description{
Code after a top-level \code{\link[=return]{return()}} or \code{\link[=stop]{stop()}} can't be reached; typically
this is vestigial code left after refactoring or sandboxing code, which is
fine for exploration, but shouldn't ultimately be checked in. Comments
meant for posterity should be placed \emph{before} the final \code{return()}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=readability_linters]{readability}
}
