% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_import_linter.R
\name{unused_import_linter}
\alias{unused_import_linter}
\title{Check that imported packages are actually used}
\usage{
unused_import_linter(except_packages = c("bit64", "data.table", "tidyverse"))
}
\arguments{
\item{except_packages}{Character vector of packages that are ignored.
These are usually attached for their side effects.}
}
\description{
Check that imported packages are actually used
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
\link[=best_practices_linters]{best_practices}, \link[=correctness_linters]{correctness}
}
