% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_lintr.R
\name{use_lintr}
\alias{use_lintr}
\title{Use lintr in your project}
\usage{
use_lintr(path = ".", type = c("minimal", "full"))
}
\arguments{
\item{path}{Path to project root, where a \code{.lintr} file should be created.
If it already exists, an error will be thrown.}
}
\value{
Path to the generated configuration, invisibly.
}
\description{
Create a minimal lintr config file as a starting point for customization
}
\examples{
\dontrun{
# use the default set of linters
lintr::use_lintr()
# or try all linters
lintr::use_lintr(type = "full")

# then
lintr::lint_dir()
}
}
