% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_utils.R
\name{xml_nodes_to_lints}
\alias{xml_nodes_to_lints}
\title{Convert an XML node or nodeset into a Lint}
\usage{
xml_nodes_to_lints(
  xml,
  source_expression,
  lint_message,
  type = c("style", "warning", "error"),
  match_after_end = FALSE
)
}
\arguments{
\item{xml}{An \code{xml_node} object (to generate one \code{Lint}) or an
\code{xml_nodeset} object (to generate several \code{Lint}s), e.g. as returned by
\code{\link[xml2:xml_find_all]{xml2::xml_find_all()}} or \code{\link[xml2:xml_find_all]{xml2::xml_find_first()}}.}

\item{source_expression}{A source expression object, e.g. as
returned typically by \code{\link[=lint]{lint()}}, or more generally
by \code{\link[=get_source_expressions]{get_source_expressions()}}.}

\item{lint_message}{The message to be included as the \code{message}
to the \code{Lint} object. If \code{lint_message} is a \code{function},
this function is first applied to \code{xml} (so it should be a
function taking an \code{xml_node} as input and must produce a
length-1 character as output).}

\item{type}{type of lint.}

\item{match_after_end}{Logical, default \code{FALSE}. If \code{TRUE},
The output \code{column_number} and \code{ranges[2L]} are set to \emph{after} the matched
symbol in \code{xml}. This can be convenient for setting the source marker
to land the editor's cursor exactly to the spot where corrected code can be
entered. For example, in \code{\link[=T_and_F_symbol_linter]{T_and_F_symbol_linter()}}, in code using \code{T}, the
cursor lands \emph{after} \code{T} so that \code{RUE} and be entered without further adjustment.}
}
\value{
For \code{xml_node}s, a \code{lint}. For \code{xml_nodeset}s, \code{lints} (a list of \code{lint}s).
}
\description{
Convenience function for converting nodes matched by XPath-based
linter logic into a \code{\link[=Lint]{Lint()}} object to return.
}
