% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoda_test_linter.R
\name{yoda_test_linter}
\alias{yoda_test_linter}
\title{Block obvious "yoda tests"}
\usage{
yoda_test_linter()
}
\description{
Yoda tests (\href{https://goto.google.com/c-readability-advice#avoid-yoda-style-comparisons}{go/c-readability-advice#avoid-yoda-style-comparisons}) test
like \verb{(expected, actual)} instead of the more common \verb{(actual, expected)}.
This is not always possible to detect statically; this linter focuses on
the simple case of testing an expression against a literal value, e.g.
\verb{(1L, foo(x))} should be \verb{(foo(x), 1L)}.
}
\seealso{
\link{linters} for a complete list of linters available in lintr.
}
\section{Tags}{
No tags are given.
}
